% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_plot.R
\name{add_eirpr_seasonal_profile}
\alias{add_eirpr_seasonal_profile}
\title{Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.}
\usage{
add_eirpr_seasonal_profile(ix, xds_obj, clrs)
}
\arguments{
\item{ix}{the subset of with orbits to plot}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clrs}{a \link{character} vector of colors}
}
\description{
Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.
}
