% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{add_fit_spline_ty}
\alias{add_fit_spline_ty}
\title{Change Spline \eqn{t,y} Values}
\usage{
add_fit_spline_ty(new_y, new_t, xds_obj)
}
\arguments{
\item{new_y}{the new \eqn{y} interpolation point value(s)}

\item{new_t}{the new \eqn{t} interpolation point value(s)}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Change Spline \eqn{t,y} Values
}
