% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{add_orbits}
\alias{add_orbits}
\title{Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.}
\usage{
add_orbits(i, pars, clr)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{clrs}{a [character] vector of colors}

\item{llty}{a [list]}
}
\description{
Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.
}
