% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-season.R
\name{approx_phase}
\alias{approx_phase}
\title{Compute the observed phase}
\usage{
approx_phase(pfpr, jdate)
}
\arguments{
\item{pfpr}{a \emph{Pf}PR time series}

\item{jdate}{the corresponding dates}
}
\value{
a list with the mean peak and the values
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
a spline function
}
