% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dts-step-ahead.R
\name{ar_compare}
\alias{ar_compare}
\title{Compute the attack rate to get}
\usage{
ar_compare(ar, pr, get_pr, model, n = 1)
}
\arguments{
\item{ar}{the attack rate}

\item{pr}{next value of the PfPR}

\item{get_pr}{a function to compute PR, such as F_pr}

\item{model}{a model}

\item{n}{the number of times to iterate}
}
\value{
the sum of squared errors
}
\description{
Compute the sum of squared errors
}
