% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{bracket_lambda}
\alias{bracket_lambda}
\title{Get High/Low Values for Lambda}
\usage{
bracket_lambda(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a pair of values
}
\description{
Get High/Low Values for Lambda
}
