% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin}
\alias{burnin}
\title{Burn-in}
\usage{
burnin(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Use the defined hindcast to
run a model from the past up to the present.
Reset the initial conditions.
}
