% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{change_all_fit_spline_t}
\alias{change_all_fit_spline_t}
\title{Replace Spline \eqn{t} Values}
\usage{
change_all_fit_spline_t(new_t, xds_obj)
}
\arguments{
\item{new_t}{a new set of \eqn{t} values for the interpolation points}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Replace Spline \eqn{t} Values
}
