% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season.R
\name{change_F_season}
\alias{change_F_season}
\title{Change F_season}
\usage{
change_F_season(X, model)
}
\arguments{
\item{X}{the value of the parameters}

\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Change F_season
}
