% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season.R
\name{change_F_season_phase.Lambda}
\alias{change_F_season_phase.Lambda}
\title{Change F_season}
\usage{
\method{change_F_season_phase}{Lambda}(phase, model)
}
\arguments{
\item{phase}{the value of the parameters}

\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Change F_season
}
