% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{change_forecast_ix}
\alias{change_forecast_ix}
\title{Make a Forecast from Trusted Values}
\usage{
change_forecast_ix(xds_obj, trusted_ix, N = 3, impute_y = "subsamp")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{trusted_ix}{the index (or indices) of a node (or nodes) to replace}

\item{N}{the number of years to forecast}

\item{impute_y}{a text string to dispatch \code{impute_value}}
}
\value{
an \strong{\code{ramp.xds}} model object
}
\description{
Use trusted values from a fitted model
to generate a new set of interpolation points
}
