% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{change_forecast_ty}
\alias{change_forecast_ty}
\title{Impute the baseline}
\usage{
change_forecast_ty(xds_obj, t, y)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{t}{new \eqn{t}-values for the forecast interpolation points}

\item{y}{new \eqn{y}-values for the forecast interpolation points}
}
\value{
an \strong{\code{ramp.xds}} model object
}
\description{
Impute the a new \eqn{y} value(s) for one or more
interpolation points
}
