% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{change_ix_fit_spline_t}
\alias{change_ix_fit_spline_t}
\title{Change Spline \eqn{t} Values}
\usage{
change_ix_fit_spline_t(new_t, t_ix, xds_obj)
}
\arguments{
\item{new_t}{the new \eqn{t} interpolation point value(s)}

\item{t_ix}{the indices of the values to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Change Spline \eqn{t} Values
}
