% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{change_yy}
\alias{change_yy}
\title{Update `F_trend`}
\usage{
change_yy(yy, ix, model)
}
\arguments{
\item{yy}{the \eqn{y} value for interpolation}

\item{ix}{the index of interpolation point to replace}

\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Use the `fitting` object to update `trend_par` and
`F_trend` when the model is forced by the eir
}
