% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_setup.R
\name{check_season_par.Lambda}
\alias{check_season_par.Lambda}
\title{Check the seasonal setup}
\usage{
\method{check_season_par}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Check the seasonal setup
}
