% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{compile_fit_trend}
\alias{compile_fit_trend}
\title{Check the trend setup}
\usage{
compile_fit_trend(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Check the trend setup
}
