% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gof.R
\name{compute_gof}
\alias{compute_gof}
\title{Compute GoF}
\usage{
compute_gof(model, pfpr_ts)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{pfpr_ts}{a *Pf*PR time series}
}
\value{
sum of squared differences
}
\description{
Compute GoF
}
