% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gof.R
\name{compute_gof.sse}
\alias{compute_gof.sse}
\title{Compute GoF by LSS}
\usage{
\method{compute_gof}{sse}(model, pfpr_ts)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{pfpr_ts}{a *Pf*PR time series}
}
\value{
sum of squared differences
}
\description{
Compute GoF by LSS
}
