% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-compute_gof.R
\name{compute_gof.ts_sse}
\alias{compute_gof.ts_sse}
\title{Compute SSE}
\usage{
\method{compute_gof}{ts_sse}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
sum of squared differences
}
\description{
Computes the sum of squared errors
}
