% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_compute_gof.R
\name{compute_gof_X}
\alias{compute_gof_X}
\title{Compute the GoF for \code{X}}
\usage{
compute_gof_X(X, xds_obj, update, ix)
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{a string to dispatch}

\item{ix}{indices or a list}
}
\value{
sum of squared differences
}
\description{
Compute the GoF for \code{X}
}
