% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.R
\name{compute_impact}
\alias{compute_impact}
\title{Compute Measures of Impact}
\usage{
compute_impact(
  history,
  counterfactual,
  times = c(),
  event_date = c(),
  eval_period = 365
)
}
\arguments{
\item{history}{a model of the history: a \strong{\code{ramp.xds}} xds_obj object}

\item{counterfactual}{a model of the counterfactual: a \strong{\code{ramp.xds}} xds_obj}

\item{times}{a set of times to simulate}

\item{event_date}{the date of an event}

\item{eval_period}{the date of an event}
}
\value{
a list with measures of impact
}
\description{
Compute Measures of Impact
}
