% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_dispersal.R
\name{compute_IoD_S}
\alias{compute_IoD_S}
\title{Compute the Index of Dispersal for a seasonal pattern
\eqn{S(t)} over one year. This assumes that
\deqn{\int_0^{365} S(t)dt = 365}}
\usage{
compute_IoD_S(S)
}
\arguments{
\item{S}{a function describing a seasonal pattern}
}
\description{
If in doubt, use \link{compute_IoD_F}
}
