% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{compute_IoSD_pr}
\alias{compute_IoSD_pr}
\title{Compute the Index of Seasonal Dispersal for a PR seasonal orbit
for the \eqn{i^{th}} element of \code{eirpr$scaling}}
\usage{
compute_IoSD_pr(i, pars)
}
\arguments{
\item{i}{the index of the orbit to be evaluated}

\item{pars}{an \strong{\code{xds}} object}
}
\description{
Compute the Index of Seasonal Dispersal for a PR seasonal orbit
for the \eqn{i^{th}} element of \code{eirpr$scaling}
}
