% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{compute_IoSD_S}
\alias{compute_IoSD_S}
\title{Compute the Index of Seasonal Dispersal for a seasonal pattern
\eqn{S(t)} over one year. This assumes that
\deqn{\int_0^{365} S(t)dt = 365}}
\usage{
compute_IoSD_S(S)
}
\arguments{
\item{S}{a function describing a seasonal pattern}
}
\description{
If in doubt, use \link{compute_IoSD_F}
}
