% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{compute_threshold}
\alias{compute_threshold}
\title{Get High/Low Values for Lambda}
\usage{
compute_threshold(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a pair of values
}
\description{
Get High/Low Values for Lambda
}
