% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{crude_fit_trend.Lambda}
\alias{crude_fit_trend.Lambda}
\title{Initialize the trend}
\usage{
\method{crude_fit_trend}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
For a model forced by emergence,
pre-calibrate the forcing function so that it's
close.
}
