% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{crude_fit_trend}
\alias{crude_fit_trend}
\title{Initialize the trend}
\usage{
crude_fit_trend(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Before fitting,
pre-calibrate the forcing function so that it's
"close."
}
