% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dts_compute_gof}
\alias{dts_compute_gof}
\title{sum of squared errors}
\usage{
dts_compute_gof(ppp, data, model, F_obs, put_par, F_gof, Tmax = 365, dt = 1)
}
\arguments{
\item{ppp}{the parameters to be fitted}

\item{data}{a set of observations}

\item{model}{a model}

\item{F_obs}{a function to return the predicted observations from the model}

\item{put_par}{a function to put the parameters to be fitted}

\item{F_gof}{a function to compute the goodness of fit}

\item{Tmax}{the maximum runtime}

\item{dt}{the time step}
}
\value{
the sum of squared errors
}
\description{
Compute the sum of squared errors
}
