% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dts-step-ahead.R
\name{dts_init_by_pr}
\alias{dts_init_by_pr}
\title{Initialize a dts model such that}
\usage{
dts_init_by_pr(pr, model, get_pr, n = 365, tol = 1e-08)
}
\arguments{
\item{pr}{a PR time series}

\item{model}{a model}

\item{get_pr}{a functrion to compute PR, such as F_pr}

\item{n}{the number of times to iterate}

\item{tol}{the tolerance}
}
\value{
the model, initialized
}
\description{
Compute the sum of squared errors
}
