% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dts_maximize_gof}
\alias{dts_maximize_gof}
\title{sum of squared errors}
\usage{
dts_maximize_gof(
  data,
  model,
  F_obs,
  get_par,
  put_par,
  F_gof,
  Tmax = 365,
  dt = 1,
  tol = 1e-08
)
}
\arguments{
\item{data}{a set of observations}

\item{model}{a model}

\item{F_obs}{a function to return the predicted observations from the model}

\item{get_par}{a function to get the parameters to be fitted}

\item{put_par}{a function to put the parameters to be fitted}

\item{F_gof}{a function to compute the goodness of fit}

\item{Tmax}{the maximum runtime}

\item{dt}{the time step}
}
\value{
the sum of squared errors
}
\description{
Compute the sum of squared errors
}
