% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dts-step-ahead.R
\name{dts_pr2ar_step}
\alias{dts_pr2ar_step}
\title{Compute the attack rates for a set of strata for a step-ahead algorithm}
\usage{
dts_pr2ar_step(pr, model, get_pr, n = 365, tol = 1e-08)
}
\arguments{
\item{pr}{next value of the PfPR}

\item{model}{a model}

\item{get_pr}{a function to compute PR, such as F_pr}

\item{n}{the number of times to iterate}

\item{tol}{the tolerance}
}
\value{
the sum of squared errors
}
\description{
Compute the sum of squared errors
}
