% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dts-step-ahead.R
\name{dts_pr2ar_ts}
\alias{dts_pr2ar_ts}
\title{Compute the attack rate to get}
\usage{
dts_pr2ar_ts(pr, model, get_pr, tol = 1e-08)
}
\arguments{
\item{pr}{a PR time series}

\item{model}{a model}

\item{get_pr}{a function to compute PR, such as F_pr}

\item{tol}{the tolerance}
}
\value{
a vector of attack rates, as a time series
}
\description{
Compute the sum of squared errors
}
