% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_dts.R
\name{dts_scaling}
\alias{dts_scaling}
\title{Scaling for Malaria Metrics}
\usage{
dts_scaling(xds_obj, N = 25, rbr = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}

\item{rbr}{the relative biting rate for a population stratum}
}
\value{
\strong{\code{xds}} xds_obj object
}
\description{
This function dispatches on \code{class(xds_obj$forced_by)}
}
