% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{effect_sizes_set_zap}
\alias{effect_sizes_set_zap}
\title{Set the zap parameters}
\usage{
effect_sizes_set_zap(X, model, irs_rounds, itn_rounds)
}
\arguments{
\item{X}{a set of parameters to be fit}

\item{model}{a **`ramp.xds`** model object}

\item{irs_rounds}{a list to configure irs_multiround}

\item{itn_rounds}{a list to configure itn_multiround}
}
\description{
Set the zap parameters
to fit effect sizes
}
