% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{eirpr_seasonal_profile}
\alias{eirpr_seasonal_profile}
\title{Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.}
\usage{
eirpr_seasonal_profile(ix, pars, clrs)
}
\arguments{
\item{ix}{the subset of with orbits to plot}

\item{pars}{an **`xds`** object}

\item{clrs}{a [character] vector of colors}
}
\description{
Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.
}
