% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_baseline.R
\name{estimate_effect_size.bednet}
\alias{estimate_effect_size.bednet}
\title{Estimate the effect size for a single round of mass bed net distribution}
\usage{
\method{estimate_effect_size}{bednet}(event_ix, model, pfpr_ts, jdates)
}
\arguments{
\item{event_ix}{the index of the event to evaluate}

\item{model}{a **`ramp.xds`** model object}

\item{pfpr_ts}{the PR observed}

\item{jdates}{the julian dates of the observations}
}
\description{
This function fits the effect size for a bed net distribution.
}
