% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{estimate_effect_size.bednets}
\alias{estimate_effect_size.bednets}
\title{Estimate the effect size for a single round of mass bed net distribution}
\usage{
\method{estimate_effect_size}{bednets}(model, data, times, event)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{event}{relevant details of an event to assess}
}
\description{
This function fits the effect size for a bed net distribution.
}
