% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{estimate_effect_size.irs}
\alias{estimate_effect_size.irs}
\title{Estimate the effect size for a single round of IRS}
\usage{
\method{estimate_effect_size}{irs}(model, data, times, event)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{event}{relevant details of an event to assess}
}
\description{
This function fits effect sizes for vector-based interventions. Each
intervention in a list includes a temporal profile describing how the
*effect* of the intervention wanes over time and the inception time.
Each model of intervention -- IRS or ITNs -- has its own sub-model that
describes how bionomic parameters are modified by vector control, given
contact and the waning effects.
}
