% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_baseline.R
\name{estimate_effect_size}
\alias{estimate_effect_size}
\title{Estimate the effect size for a single round of IRS}
\usage{
estimate_effect_size(event_ix, model, pfpr_ts, jdates)
}
\arguments{
\item{event_ix}{the index of the event to evaluate}

\item{model}{a **`ramp.xds`** model object}

\item{pfpr_ts}{the PR observed}

\item{jdates}{the julian dates of the observations}
}
\description{
This function fits effect sizes for vector-based interventions. Each
intervention in a list includes a temporal profile describing how the
*effect* of the intervention wanes over time and the inception time.
Each model of intervention -- IRS or ITNs -- has its own sub-model that
describes how bionomic parameters are modified by vector control, given
contact and the waning effects.
}
