% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{estimate_effect_sizes}
\alias{estimate_effect_sizes}
\title{Estimate the effect sizes of vector control}
\usage{
estimate_effect_sizes(
  model,
  base_ix,
  data,
  times,
  irs_rounds = list(),
  itn_rounds = list()
)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{base_ix}{counterfactual baseline knots}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{irs_rounds}{a list to configure irs_multiround}

\item{itn_rounds}{a list to configure itn_multiround}
}
\description{
This function fits effect sizes for vector-based interventions. Each
intervention in a list includes a temporal profile describing how the
*effect* of the intervention wanes over time and the inception time.
Each model of intervention -- IRS or ITNs -- has its own sub-model that
describes how bionomic parameters are modified by vector control, given
contact and the waning effects.
}
