% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend-utils.R
\name{event_chop_spline_t}
\alias{event_chop_spline_t}
\title{Adjust Spacing}
\usage{
event_chop_spline_t(xds_obj, b4 = 30, fu = 210, mngp = 90, mxgp = 500)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{b4}{days before intervention}

\item{fu}{days after intervention}

\item{mngp}{minimum gap}

\item{mxgp}{the maximum gap}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Adjust Spacing
}
