% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_setup.R
\name{F_pr}
\alias{F_pr}
\title{Rebuild Forcing Functions}
\usage{
F_pr(tm, ts_obj, bday = 0)
}
\value{
an \strong{\code{xds}} object
}
\description{
Rebuild forcing functions on the EIR object using \code{make_function}:
\itemize{
\item \code{F_season} is made from \code{season_par}
\item \code{F_trend} is made from \code{trend_par}
\item \code{F_shock} is made from \code{shock_par}
}
}
\keyword{internal}
