% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{F_sse}
\alias{F_sse}
\title{Sum of squared errors. Works as F_gof in xde_maximize_gof}
\usage{
F_sse(obs, pred)
}
\arguments{
\item{obs}{a set of observations}

\item{pred}{a set of predicted values for the observations}
}
\value{
the sum of squared errors
}
\description{
Compute the sum of squared errors
}
