% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_contact.R
\name{fit_bednet_contact}
\alias{fit_bednet_contact}
\title{Fit bed net coverage}
\usage{
fit_bednet_contact(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{options}{setup options for the bed net rounds indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter describing
bed net coverage
}
