% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bednet_coverage.R
\name{fit_bednet_coverage}
\alias{fit_bednet_coverage}
\title{Fit bed net coverage}
\usage{
fit_bednet_coverage(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter describing
bed net coverage
}
