% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_shock.R
\name{fit_bednet_shock}
\alias{fit_bednet_shock}
\title{Fit Bednet Shock}
\usage{
fit_bednet_shock(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{options}{setup options for the bednet rounds indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter(s) describing
the scale of bednet effect sizes.

This should not be used
}
