% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_eir.R
\name{fit_EIR_phase_sin_season}
\alias{fit_EIR_phase_sin_season}
\title{Fit the phase for a time series}
\usage{
fit_EIR_phase_sin_season(data, times, model, Fp = NULL)
}
\arguments{
\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` object}

\item{Fp}{initial parameters for seasonality}
}
\value{
an `xds` object
}
\description{
For a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
