% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_eir.R
\name{fit_EIR_sin_season}
\alias{fit_EIR_sin_season}
\title{Fit the phase for a time series}
\usage{
fit_EIR_sin_season(data, times, model, Fpar = NULL)
}
\arguments{
\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` object}

\item{Fpar}{parameters from `makepar_F_sin`}
}
\value{
an `xds` object
}
\description{
For a time series \eqn{X,} compute the
phase, the time of the year when there is a peak
}
