% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_irs_coverage.R
\name{fit_irs_coverage}
\alias{fit_irs_coverage}
\title{Fit irs coverage}
\usage{
fit_irs_coverage(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter(s) describing
irs coverage
}
