% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs-d50.R
\name{fit_irs_d50}
\alias{fit_irs_d50}
\title{Fit IRS shock_d50}
\usage{
fit_irs_d50(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{options}{setup options for the irs rounds indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter(s) describing
the scale of IRS effect d50s.

This should not be used
}
