% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs-dshape.R
\name{fit_irs_dshape}
\alias{fit_irs_dshape}
\title{IRS Effects: Fit Shape}
\usage{
fit_irs_dshape(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{options}{setup options for the irs rounds indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter(s) \code{d_shape} describing
the shape of the response timeline
}
