% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs_shock.R
\name{fit_irs_shock}
\alias{fit_irs_shock}
\title{Fit IRS Shock}
\usage{
fit_irs_shock(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{options}{setup options for the irs rounds indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameters describing
the response timeline to IRS. This fits three parameters:
\itemize{
\item \code{size} controls the maximum effect size of the shock
\item \code{d_50} days after the shock when the effect size reaches half of the maximum
\item \code{d_shape} a shape parameter
}
}
