% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs_shock-size.R
\name{fit_irs_shock_size}
\alias{fit_irs_shock_size}
\title{Fit IRS shock_size}
\usage{
fit_irs_shock_size(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{options}{setup options for the irs rounds indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
As part of a dynamical time series
analysis, fit the parameter(s) describing
the scale of IRS effect sizes.

This should not be used
}
