% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{fit_Lambda_amplitude_sin_season}
\alias{fit_Lambda_amplitude_sin_season}
\title{Fit the amplitude for a time series}
\usage{
fit_Lambda_amplitude_sin_season(data, times, model, Fpar = NULL)
}
\arguments{
\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` model}

\item{Fpar}{parameters from `makepar_F_sin`}
}
\value{
an `xds` object
}
\description{
For a time series \eqn{X,} compute the
phase, the time of the year when there is a peak `dog`
}
