% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mean.R
\name{fit_mean}
\alias{fit_mean}
\title{Fit a seasonal pattern}
\usage{
fit_mean(pfpr_ts, jdates, model)
}
\arguments{
\item{pfpr_ts}{a PfPR time series}

\item{jdates}{the times of the observations}

\item{model}{an **`xds`** model object}
}
\value{
a list with the mean peak and the values
}
\description{
For a *Pf*PR time series, \eqn{x}, fit
a spline function
}
