% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mean_forcing.R
\name{fit_mean_forcing}
\alias{fit_mean_forcing}
\title{Fit mean forcing}
\usage{
fit_mean_forcing(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
\code{xds_obj}
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
the mean forcing parameter
}
