% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model to data}
\usage{
fit_model(xds_obj, update, ix)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{dispatch string(s)}

\item{ix}{indices, for model fitting}
}
\value{
a list with the mean peak and the values
}
\description{
Fit a model to data
}
