% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season.R
\name{fit_phase}
\alias{fit_phase}
\title{Fit the phase}
\usage{
fit_phase(pfpr_ts, jdates, model, n = 2)
}
\arguments{
\item{pfpr_ts}{a PfPR time series}

\item{jdates}{the times of the observations}

\item{model}{an `xds` model}

\item{n}{number of times to iterate}
}
\value{
a list with the mean peak and the values
}
\description{
For a *Pf*PR time series, \eqn{x}, fit
the phase for a seasonal pattern. The algorithm works
by choosing an initial mesh, around 46 days, and then
shrinking the mesh around the value that maximizes
the goodness of fit. The \eqn{n} parameter determines
how many times the process is iterated. For \eqn{n=2},
the mesh is
}
