% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{fit_phase_sin_season}
\alias{fit_phase_sin_season}
\title{Fit the phase for a time series}
\usage{
fit_phase_sin_season(data, times, model, Fp = NULL)
}
\arguments{
\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` model}

\item{Fp}{parameters for}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series \eqn{X,} compute the
phase, the time of the year when there is a peak
}
