% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_phase.R
\name{fit_season_phase_alt}
\alias{fit_season_phase_alt}
\title{Fit the phase}
\usage{
fit_season_phase_alt(xds_obj, n = 2)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{n}{number of times to iterate}
}
\value{
a list with the mean peak and the values
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
the phase for a seasonal pattern. The algorithm works
by choosing an initial mesh, around 46 days, and then
shrinking the mesh around the value that maximizes
the goodness of fit. The \eqn{n} parameter determines
how many times the process is iterated. For \eqn{n=2},
the mesh is
}
