% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_shock.R
\name{fit_shock}
\alias{fit_shock}
\title{Fit interannual variability using splines}
\usage{
fit_shock(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{options}{a list to configure the indices}
}
\value{
a list with the mean peak and the values
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
a spline function. To fit only a subset of the interpolation
points, define \code{options$shock_ix}
}
