% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{fit_spline}
\alias{fit_spline}
\title{Fit the amplitude for a time series}
\usage{
fit_spline(data, times, model, splinef = 2)
}
\arguments{
\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` model}

\item{splinef}{which type of spline}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series \eqn{X,} compute the
inter-annual variability
}
