% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{fit_splinef}
\alias{fit_splinef}
\title{Fit the amplitude for a time series}
\usage{
fit_splinef(N, data, times, model)
}
\arguments{
\item{N}{the number of knots}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` model}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series \eqn{X,} compute the
inter-annual variability
}
