% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Fit interannual variability using splines}
\usage{
fit_trend(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{options}{a list to configure the indices}
}
\value{
an \code{xds} xds_obj
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
a spline function. To fit only a subset of the interpolation
points, define \code{options$trend_ix}
}
