% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend_spline}
\alias{fit_trend_spline}
\title{Fit interannual variability using splines}
\usage{
fit_trend_spline(pfpr_ts, jdates, model, ix = c())
}
\arguments{
\item{pfpr_ts}{a PfPR time series}

\item{jdates}{the times of the observations}

\item{model}{an `xds` model}

\item{ix}{the interpolation points to fit}
}
\value{
a list with the mean peak and the values
}
\description{
For a *Pf*PR time series, \eqn{x}, fit
a spline function
}
