% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-change_trend.R
\name{fitting_change_spline_ty}
\alias{fitting_change_spline_ty}
\title{Change Spline \eqn{t,y} Values}
\usage{
fitting_change_spline_ty(new_y, new_t, ix, xds_obj)
}
\arguments{
\item{new_y}{the new \eqn{y} interpolation point value(s)}

\item{new_t}{the new \eqn{t} interpolation point value(s)}

\item{ix}{the indices of the values to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Change Spline \eqn{t,y} Values
}
