% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-change_trend.R
\name{fitting_change_spline_y}
\alias{fitting_change_spline_y}
\title{Change Spline \eqn{y} Values}
\usage{
fitting_change_spline_y(new_y, y_ix, xds_obj)
}
\arguments{
\item{new_y}{the new \eqn{y} interpolation point value(s)}

\item{y_ix}{the indices of the values to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Change Spline \eqn{y} Values
}
