% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-change_trend.R
\name{fitting_get_spline_ty}
\alias{fitting_get_spline_ty}
\title{Get Spline \eqn{t, y} Values}
\usage{
fitting_get_spline_ty(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Get Spline \eqn{t, y} Values
}
