% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-change_trend.R
\name{fitting_replace_spline_y}
\alias{fitting_replace_spline_y}
\title{Replace Spline \eqn{y} Values}
\usage{
fitting_replace_spline_y(new_y, xds_obj)
}
\arguments{
\item{new_y}{a new set of \eqn{y} values for the interpolation points}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Replace Spline \eqn{y} Values
}
