% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_spline}
\alias{forecast_spline}
\title{Add a "forecast"}
\usage{
forecast_spline(spline_par, N = 5)
}
\arguments{
\item{spline_par}{parameters to set up a spline function}

\item{N}{the number of years ahead to forecast}
}
\description{
Use the nodes of a trend function, defined as a spline, to
"forecast" the future by sub-sampling from the knots
}
