% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_spline_Lambda}
\alias{forecast_spline_Lambda}
\title{Add a "forecast"}
\usage{
forecast_spline_Lambda(model, N = 5, s = 1)
}
\arguments{
\item{model}{a list that defines an `ramp.xds` model (*e.g.*,  generated by `xde_setup()`)}

\item{N}{the number of years ahead to forecast}

\item{s}{the species to forecast}
}
\description{
Use the nodes of a trend function, defined as a spline, to
"forecast" the future by sub-sampling from the knots
}
