% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{gam_forecast}
\alias{gam_forecast}
\title{Fit and draw}
\usage{
gam_forecast(y, N)
}
\arguments{
\item{y}{values to fit}

\item{N}{the number of observations}
}
\value{
random draws
}
\description{
Fit and draw
}
