% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{gam_sample}
\alias{gam_sample}
\title{Sample from a fitted gamma}
\usage{
gam_sample(yy, N)
}
\arguments{
\item{yy}{the knots}

\item{N}{the number of years ahead to forecast}
}
\description{
Use the nodes of a trend function, defined as a spline, to
"forecast" the future by sub-sampling from a fitted gamma
}
