% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{get_fit_trend}
\alias{get_fit_trend}
\title{Get Spline \eqn{t, y} Values}
\usage{
get_fit_trend(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Get Spline \eqn{t, y} Values
}
