% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_amplitude.R
\name{get_init_X.amplitude}
\alias{get_init_X.amplitude}
\title{Get Initial Values for Parameters}
\usage{
\method{get_init_X}{amplitude}(xds_obj, update = "amplitude", ix = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
