% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bednet_coverage.R
\name{get_init_X.bednet_coverage}
\alias{get_init_X.bednet_coverage}
\title{Get initial X: Bed net coverage}
\usage{
\method{get_init_X}{bednet_coverage}(xds_obj, update = "bednet_coverage", ix = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
bednet coverage, as \link{numeric}
}
\description{
Get initial X: Bed net coverage
}
