% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet-dshape.R
\name{get_init_X.bednet_dshape}
\alias{get_init_X.bednet_dshape}
\title{Get initial X: bednet dshape}
\usage{
\method{get_init_X}{bednet_dshape}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
bednet dshape levels
}
\description{
Get initial X: bednet dshape
}
