% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_shock.R
\name{get_init_X.bednet_shock}
\alias{get_init_X.bednet_shock}
\title{Get initial X: bednet shock}
\usage{
\method{get_init_X}{bednet_shock}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
bednet shock levels
}
\description{
Get initial X: bednet shock
}
