% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_shock-d50.R
\name{get_init_X.bednet_shock_d50}
\alias{get_init_X.bednet_shock_d50}
\title{Get initial X: bednet d50}
\usage{
\method{get_init_X}{bednet_shock_d50}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
bednet d50 levels
}
\description{
Get initial X: bednet d50
}
