% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_irs_contact.R
\name{get_init_X.irs_contact}
\alias{get_init_X.irs_contact}
\title{Get initial X: IRS contact}
\usage{
\method{get_init_X}{irs_contact}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
IRS contact levels
}
\description{
Get initial X: IRS contact
}
