% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_irs_coverage.R
\name{get_init_X.irs_coverage}
\alias{get_init_X.irs_coverage}
\title{Get initial X: IRS coverage}
\usage{
\method{get_init_X}{irs_coverage}(xds_obj, update = "irs_coverage", ix = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
IRS coverage levels
}
\description{
Get initial X: IRS coverage
}
