% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs-dshape.R
\name{get_init_X.irs_dshape}
\alias{get_init_X.irs_dshape}
\title{Get initial X: IRS dshape}
\usage{
\method{get_init_X}{irs_dshape}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
IRS dshape levels
}
\description{
Get initial X: IRS dshape
}
