% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs_shock.R
\name{get_init_X.irs_shock}
\alias{get_init_X.irs_shock}
\title{Get initial X: IRS shock}
\usage{
\method{get_init_X}{irs_shock}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
IRS shock levels
}
\description{
Get initial X: IRS shock
}
