% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs_shock-size.R
\name{get_init_X.irs_shock_size}
\alias{get_init_X.irs_shock_size}
\title{Get initial X: IRS shock_size}
\usage{
\method{get_init_X}{irs_shock_size}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
IRS shock_size levels
}
\description{
Get initial X: IRS shock_size
}
