% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_init_X.multifit}
\alias{get_init_X.multifit}
\title{Compute the GoF for \code{X}}
\usage{
\method{get_init_X}{multifit}(xds_obj, update, ix)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
sum of squared differences
}
\description{
Compute the GoF for \code{X}
}
