% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_pw.R
\name{get_init_X.pw}
\alias{get_init_X.pw}
\title{Get Initial Values for Parameters}
\usage{
\method{get_init_X}{pw}(xds_obj, update = "pw", ix = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
