% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_compute_gof.R
\name{get_init_X}
\alias{get_init_X}
\title{Get Initial Values for Parameters}
\usage{
get_init_X(xds_obj, update, ix = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
a list with the mean peak and the values
}
\description{
Get Initial Values for Parameters
}
