% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{get_init_X.trend}
\alias{get_init_X.trend}
\title{Get Initial Values for Parameters}
\usage{
\method{get_init_X}{trend}(xds_obj, update = "trend", ix = c())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}

\item{ix}{the index}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
