% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_contact.R
\name{get_limits_X.bednet_contact}
\alias{get_limits_X.bednet_contact}
\title{Get limits for IRS coverage parameters}
\usage{
\method{get_limits_X}{bednet_contact}(xds_obj, feature = "bednet_contact")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
bounds on coverage, as \link{numeric}
}
\description{
Get limits for IRS coverage parameters
}
