% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bednet_coverage.R
\name{get_limits_X.bednet_coverage}
\alias{get_limits_X.bednet_coverage}
\title{Get limits for IRS coverage parameters}
\usage{
\method{get_limits_X}{bednet_coverage}(xds_obj, update = "bednet_coverage")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}
}
\value{
bounds on coverage, as \link{numeric}
}
\description{
Get limits for IRS coverage parameters
}
