% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet-dshape.R
\name{get_limits_X.bednet_dshape}
\alias{get_limits_X.bednet_dshape}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{bednet_dshape}(xds_obj, feature = "bednet_dshape")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
