% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_shock-size.R
\name{get_limits_X.bednet_shock_size}
\alias{get_limits_X.bednet_shock_size}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{bednet_shock_size}(xds_obj, feature = "bednet_shock_size")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
