% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_bottom.R
\name{get_limits_X.bottom}
\alias{get_limits_X.bottom}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{bottom}(xds_obj, feature = "bottom")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
