% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_irs_coverage.R
\name{get_limits_X.irs_coverage}
\alias{get_limits_X.irs_coverage}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{irs_coverage}(xds_obj, feature = "irs_coverage")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
