% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-season-pw.R
\name{get_limits_X.pw}
\alias{get_limits_X.pw}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{pw}(xds_obj, feature = "pw")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
