% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-compute_gof.R
\name{get_limits_X}
\alias{get_limits_X}
\title{Get Limits}
\usage{
get_limits_X(xds_obj, feature)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a list with the mean peak and the values
}
\description{
Get Limits
}
