% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_shock.R
\name{get_limits_X.shock}
\alias{get_limits_X.shock}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{shock}(xds_obj, feature)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
