% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predicted_pr.R
\name{get_predicted_pr.rdt}
\alias{get_predicted_pr.rdt}
\title{Title}
\usage{
\method{get_predicted_pr}{rdt}(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Title
}
