% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{get_spline_yy.full}
\alias{get_spline_yy.full}
\title{Modify the Baseline}
\usage{
\method{get_spline_yy}{full}(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{counterfactual}{}
}
\value{
a **`ramp.xds`** model
}
\description{
This function modifies the parameters
in a model to set up a counter
}
