% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{get_spline_yy}
\alias{get_spline_yy}
\title{Modify the Baseline}
\usage{
get_spline_yy(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a [vector]
}
\description{
This function modifies the parameters
in a model to set up a counter
}
