% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend-get_ty.R
\name{get_trend_ty.last}
\alias{get_trend_ty.last}
\title{Get interpolation points}
\usage{
\method{get_trend_ty}{last}(ix, xds_obj, trust_ty)
}
\arguments{
\item{ix}{index or indices of interpolation points}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{trust_ty}{a text string to dispatch \code{get_trend_ty}}
}
\value{
Interpolation points, \eqn{t,y}, as a list
}
\description{
Return the last interpolation point
}
