% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trusted_ty.R
\name{get_trusted_ty.last}
\alias{get_trusted_ty.last}
\title{Get interpolation points}
\usage{
\method{get_trusted_ty}{last}(ix, model, trust_ty)
}
\arguments{
\item{ix}{index or indices of interpolation points}

\item{model}{a **`ramp.xds`** model object}

\item{trust_ty}{a text string to dispatch `get_trusted_ty`}
}
\value{
Interpolation points, \eqn{t,y}, as a list
}
\description{
Return the last interpolation point
}
