% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trusted_ty.R
\name{get_trusted_ty}
\alias{get_trusted_ty}
\title{Get a set of trusted interpolation points}
\usage{
get_trusted_ty(ix, model, trust_ty)
}
\arguments{
\item{ix}{index or indices of interpolation points}

\item{model}{a **`ramp.xds`** model object}

\item{trust_ty}{a text string to dispatch `get_trusted_ty`}
}
\value{
Interpolation points, \eqn{t,y}, as a list
}
\description{
Get a set of trusted interpolation points
}
