% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_get_ty.R
\name{get_ty.first}
\alias{get_ty.first}
\title{Get interpolation points}
\usage{
\method{get_ty}{first}(ix, model, trust_ty)
}
\arguments{
\item{ix}{index or indices of interpolation points}

\item{model}{a \strong{\code{ramp.xds}} model object}

\item{trust_ty}{a text string to dispatch \code{get_ty}}
}
\value{
Interpolation points, \eqn{t,y}, as a list
}
\description{
Return the first interpolation point
}
