% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_get_ty.R
\name{get_ty}
\alias{get_ty}
\title{Get a set of trusted interpolation points}
\usage{
get_ty(ix, model, trust_ty)
}
\arguments{
\item{ix}{index or indices of interpolation points}

\item{model}{a \strong{\code{ramp.xds}} model object}

\item{trust_ty}{a text string to dispatch \code{get_ty}}
}
\value{
Interpolation points, \eqn{t,y}, as a list
}
\description{
Get a set of trusted interpolation points
}
