% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_get_ty.R
\name{get_ty.tix}
\alias{get_ty.tix}
\title{Get interpolation points}
\usage{
\method{get_ty}{tix}(ix, xds_obj, trust_ty)
}
\arguments{
\item{ix}{index or indices of interpolation points}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{trust_ty}{a text string to dispatch \code{get_ty}}
}
\value{
Interpolation points, \eqn{t,y}, as a list
}
\description{
The \code{ix} indicates the interpolation
points that were not modified by control. The optional
argument \code{opts$t} should be the time of an event, so
the interpolation points are before the event and
unmodified by control.
}
