% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{get_yix_after_bednet_round}
\alias{get_yix_after_bednet_round}
\title{Time Since Event}
\usage{
get_yix_after_bednet_round(round_ix, xds_obj, N = 1)
}
\arguments{
\item{round_ix}{the index of the IRS round}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{the number of y indices}
}
\value{
numeric
}
\description{
Compute the time elapsed
since the last vector control event
for all the spline \eqn{t}-values
}
