% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{hindcast_ty}
\alias{hindcast_ty}
\title{Hindcast a Baseline}
\usage{
hindcast_ty(model, ix = c())
}
\arguments{
\item{model}{a \strong{\code{ramp.xds}} model object}

\item{ix}{(optional) indices for \link{impute_value}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This function adds N interpolation
points to the spline parameters \emph{before} an
observation period for burn-in. The \eqn{t}
values are set at yearly intervals, and the
\eqn{y} values that serve as a basis
}
