% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline.R
\name{impute_baseline.gam}
\alias{impute_baseline.gam}
\title{Baseline with gamma predictions}
\usage{
\method{impute_baseline}{gam}(name, model, ix)
}
\arguments{
\item{name}{the indices for the knots to be modified}

\item{model}{a **`ramp.xds`** model}

\item{ix}{the index of the node to replace}
}
\value{
an object to dispatch [modify_baseline.ttyy]
}
\description{
This function modifies
}
