% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline.R
\name{impute_baseline.linear}
\alias{impute_baseline.linear}
\title{Create a modify Baseline}
\usage{
\method{impute_baseline}{linear}(name, model, ix)
}
\arguments{
\item{model}{a **`ramp.xds`** model}

\item{ix}{the indices for the knots to be modified}
}
\value{
an object to dispatch [modify_baseline.ttyy]
}
\description{
This function modifies
}
