% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline.R
\name{impute_baseline.max}
\alias{impute_baseline.max}
\title{Use Max for Modify Baseline}
\usage{
\method{impute_baseline}{max}(name, model, ix)
}
\arguments{
\item{name}{the indices for the knots to be modified}

\item{model}{a **`ramp.xds`** model}

\item{ix}{the index of the node to replace}
}
\value{
an object to dispatch
}
\description{
This function modifies
}
