% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline.R
\name{impute_baseline}
\alias{impute_baseline}
\title{Impute the baseline}
\usage{
impute_baseline(name, model, ix)
}
\arguments{
\item{name}{the indices for the knots to be modified}

\item{model}{a **`ramp.xds`** model}

\item{ix}{the index of the node to replace}
}
\value{
an object to dispatch
}
\description{
This function modifies
}
