% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline.R
\name{impute_baseline.subsamp}
\alias{impute_baseline.subsamp}
\title{Create a modify Baseline}
\usage{
\method{impute_baseline}{subsamp}(ix, model)
}
\arguments{
\item{ix}{the indices for the knots to be modified}

\item{model}{a **`ramp.xds`** model}
}
\value{
an object to dispatch [modify_baseline.ttyy]
}
\description{
This function modifies
}
