% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline_ty.R
\name{impute_baseline_ty}
\alias{impute_baseline_ty}
\title{Impute the baseline}
\usage{
impute_baseline_ty(ix, model, impute_ty, trust_ty)
}
\arguments{
\item{ix}{the index (or indices) of a node (or nodes) to replace}

\item{model}{a **`ramp.xds`** model}

\item{impute_ty}{a text string to dispatch `impute_baseline_ty`}

\item{trust_ty}{a text string to dispatch `get_trusted_ty`}
}
\value{
numeric
}
\description{
Impute the \eqn{n} value of one or more
interpolation points in the pair, \eqn{t,y}.
}
