% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_spline_y}
\alias{impute_spline_y}
\title{Impute the baseline}
\usage{
impute_spline_y(impute_ix, trusted_ix, xds_obj, impute_y = "mean")
}
\arguments{
\item{impute_ix}{the index (or indices) of the \eqn{y} values to replace}

\item{trusted_ix}{the index (or indices) of a node (or nodes) to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{impute_y}{a text string to dispatch \code{impute_value}}
}
\value{
an \strong{\code{ramp.xds}} model object
}
\description{
Impute the a new \eqn{y} value(s) for one or more
interpolation points
}
