% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_value.asis}
\alias{impute_value.asis}
\title{Use Mean for Modify Baseline}
\usage{
\method{impute_value}{asis}(ix, xds_obj, impute_ty, trust_ty, N = c())
}
\arguments{
\item{ix}{the index (or indices) of a node (or nodes) to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{impute_ty}{a text string to dispatch \code{impute_value}}

\item{trust_ty}{a text string to dispatch \code{get_ty}}

\item{N}{(optional) the number to return}
}
\value{
numeric
}
\description{
This function modifies
}
