% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline_ty.R
\name{impute_value.max}
\alias{impute_value.max}
\title{Use Max for Modify Baseline}
\usage{
\method{impute_value}{max}(ix, model, impute_ty, trust_ty, N = c())
}
\arguments{
\item{ix}{the index (or indices) of a node (or nodes) to replace}

\item{model}{a **`ramp.xds`** model}

\item{impute_ty}{a text string to dispatch `impute_value`}

\item{trust_ty}{a text string to dispatch `get_trusted_ty`}

\item{N}{(optional) the number to return}
}
\value{
numeric
}
\description{
This function modifies
}
