% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_value.median}
\alias{impute_value.median}
\title{Use Median for Modify Baseline}
\usage{
\method{impute_value}{median}(y, impute_y, N = 1)
}
\arguments{
\item{y}{a set of y values}

\item{impute_y}{a text string to dispatch \code{impute_value}}

\item{N}{the length of the return value}
}
\value{
numeric
}
\description{
This function modifies
}
