% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-season.R
\name{init_fit_season}
\alias{init_fit_season}
\title{Set up the seasonal pattern}
\usage{
init_fit_season(season_par)
}
\arguments{
\item{season_par}{seasonal pattern parameters}
}
\value{
a \code{sin} function parameter set
}
\description{
Set up the seasonal pattern
}
\seealso{
\code{\link[ramp.func:makepar_F_sin]{ramp.func::makepar_F_sin()}}
}
