% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-compute_gof.R
\name{ksmooth_i}
\alias{ksmooth_i}
\title{Smooth PR}
\usage{
ksmooth_i(i, pr, jdates, bwidth = 180)
}
\arguments{
\item{i}{the index of an observation}

\item{pr}{predicted PR values}

\item{jdates}{the julian dates}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
smoothed PR
}
\description{
Smooth PR
}
