% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_eirpr}
\alias{lines_eirpr}
\title{Add lines for an eir-pr scaling relationship}
\usage{
lines_eirpr(pars, clrs = "black", llty = 1)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{clrs}{a [character] vector of colors}

\item{llty}{a [list]}
}
\description{
Add lines for an eir-pr scaling relationship
}
