% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{make_baseline}
\alias{make_baseline}
\title{Modify the baseline}
\usage{
make_baseline(model, t_mod)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{t_mod}{the time points when interventions were initiated}
}
\description{
This function resets the value of knots
}
