% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{make_cf_base_max}
\alias{make_cf_base_max}
\title{Use Max for Modify Baseline}
\usage{
make_cf_base_max(ix, model)
}
\arguments{
\item{ix}{the indices for the knots to be modified}

\item{model}{a **`ramp.xds`** model}
}
\value{
an object to dispatch
}
\description{
This function modifies
}
