% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{make_cf_base_ttyy}
\alias{make_cf_base_ttyy}
\title{Pass Values for Modify Baseline}
\usage{
make_cf_base_ttyy(ix, yy)
}
\arguments{
\item{ix}{the indices for the knots to be modified}

\item{yy}{the new control points}
}
\value{
an object to dispatch [modify_baseline.ttyy]
}
\description{
This function modifies
}
