% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{make_cf_no_base}
\alias{make_cf_no_base}
\title{Use Max for No Unmodified Baseline}
\usage{
make_cf_no_base(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model}

\item{ix}{the indices for the knots to be modified}
}
\value{
an object to dispatch
}
\description{
This function modifies
}
