% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_make_fit_ix.R
\name{make_fit_ix.bednet_coverage}
\alias{make_fit_ix.bednet_coverage}
\title{Indices for Bednet Models}
\usage{
\method{make_fit_ix}{bednet_coverage}(i, update, ix_opts)
}
\arguments{
\item{i}{an index for update}

\item{update}{a list of model features to update}

\item{ix_opts}{options for setting indices}
}
\value{
indices
}
\description{
Indices for Bednet Models
}
