% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_make_fit_ix.R
\name{make_fit_ix_opts}
\alias{make_fit_ix_opts}
\title{A utility to set up indices for fitting}
\usage{
make_fit_ix_opts(patch_ix = 1, trend_ix = c(), irs_ix = 1, bednet_ix = 1)
}
\arguments{
\item{patch_ix}{index across patches}

\item{trend_ix}{index across spline interpolation points}

\item{irs_ix}{index for IRS Rounds}

\item{bednet_ix}{index for Bed Net Distribution Rounds}
}
\value{
a list
}
\description{
A utility to set up indices for fitting
}
