% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vector_control_evaluation.R
\name{make_vc_event_list}
\alias{make_vc_event_list}
\title{Make an events list}
\usage{
make_vc_event_list(model, irs_history, bednet_history)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{irs_history}{history of irs: jdate of spray round; insecticide name}

\item{bednet_history}{history of bednets: jdate of mass distribution; type bednets}
}
\value{
a modified  **`ramp.xds`** model object
}
\description{
Return an ordered list of events by time and type
}
