% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mask_data.R
\name{mask_bn_event}
\alias{mask_bn_event}
\title{Mask data around a bednet event}
\usage{
mask_bn_event(i, xds_obj, daysb4 = 0, days_after = 365)
}
\arguments{
\item{i}{the index of the event to mask}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{daysb4}{the number of days before the event to mask}

\item{days_after}{the number of days after the event to mask}
}
\description{
Mask data around a bednet event
}
