% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mask_data.R
\name{mask_events}
\alias{mask_events}
\title{Mask data around events}
\usage{
mask_events(xds_obj, daysb4 = 0, days_after = 365)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{daysb4}{the number of days before the event to mask}

\item{days_after}{the number of days after the event to mask}
}
\value{
an \strong{\code{xds_obj}}
}
\description{
Mask the data around events, parameters control
the number of days before (negative values are OK) and
after.
}
