% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{mean_phase_peak}
\alias{mean_phase_peak}
\title{Compute the phase of the peak}
\usage{
mean_phase_peak(t, X, window = 180)
}
\arguments{
\item{t}{the times}

\item{X}{the data}

\item{window}{days around t}
}
\description{
For a time series \eqn{X,} compute the
phase, the time of the year when there is a peak
}
