% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{modify_baseline.func}
\alias{modify_baseline.func}
\title{Create a modify Baseline}
\usage{
\method{modify_baseline}{func}(model, counterfactual)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{counterfactual}{a list}
}
\value{
a function object
}
\description{
This function modifies
}
