% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_baseline.R
\name{modify_baseline}
\alias{modify_baseline}
\title{Modify the Baseline}
\usage{
modify_baseline(model, counterfactual)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{counterfactual}{a list}
}
\value{
a **`ramp.xds`** model
}
\description{
This function modifies the parameters
in a model to set up a counter
}
