% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-modify_vector.R
\name{modify_vector_X.NULL}
\alias{modify_vector_X.NULL}
\title{Replace Values in a List}
\usage{
\method{modify_vector_X}{`NULL`}(V, ixV, X, ixX)
}
\arguments{
\item{V}{the V to modify}

\item{ixV}{indices to be replaced}

\item{X}{the new values}

\item{ixX}{indices to be replaced}
}
\value{
the modified V
}
\description{
If \code{ix} is NULL, return \code{X}
}
\examples{

modify_vector_X(6:15, c(), 1:10)

}
