% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-modify_vector_X.R
\name{modify_vector_X.numeric}
\alias{modify_vector_X.numeric}
\title{Replace Values in a List}
\usage{
\method{modify_vector_X}{numeric}(V, ixV, X, ixX)
}
\arguments{
\item{V}{the V to modify}

\item{ixV}{indices to be replaced}

\item{X}{the new values}

\item{ixX}{indices to be replaced}
}
\value{
the modified V
}
\description{
Replace Values in a List
}
\examples{

modify_vector_X(6, c(1,7), 1:10)

}
