% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-modify_vector_X.R
\name{modify_vector_X}
\alias{modify_vector_X}
\title{Replace}
\usage{
modify_vector_X(V, ixV, X, ixX)
}
\arguments{
\item{V}{the V to modify}

\item{ixV}{indices to be replaced}

\item{X}{the new values}

\item{ixX}{indices to be replaced}
}
\value{
the modified V
}
\description{
Replace some values in a V
with some values in X
}
