% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{norm_trend.Lambda}
\alias{norm_trend.Lambda}
\title{Normalize Trend}
\usage{
\method{norm_trend}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a number
}
\description{
Normalize Trend
}
