% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{norm_trend}
\alias{norm_trend}
\title{Normalize Trend}
\usage{
norm_trend(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a number
}
\description{
Adjust the y-values for the
interpolation point and reset the mean
forcing parameter such that the product
of the seasonal pattern and the trend over
an observation period is one.
}
