% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_eirpr}
\alias{plot_eirpr}
\title{Plot the eir-pr scaling relationship}
\usage{
plot_eirpr(pars, clrs = "black", llty = 1)
}
\arguments{
\item{pars}{a [list] *e.g.*,  generated by `ramp.xds::xde_setup()` that defines a model in `ramp.xds`}

\item{clrs}{a [character] vector of colors}

\item{llty}{a [list]}
}
\description{
Plot the eir-pr scaling relationship
}
