% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2history.R
\name{pr2history_cbmc}
\alias{pr2history_cbmc}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
pr2history_cbmc(xds_obj, twice = FALSE)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{twice}{cycle through a second time}
}
\value{
the \strong{\code{ramp.xds}} model object, fitted to a time series.
The state at the end is saved as \code{xds_obj$history}
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
