% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{pr2Lambda}
\alias{pr2Lambda}
\title{Using the eirpr matrix and a RM model, convert pr to Lambda}
\usage{
pr2Lambda(pr, model, constrain = TRUE)
}
\arguments{
\item{pr}{a vector}

\item{model}{a list that defines an `ramp.xde` model (*e.g.*,  generated by `xde_setup()`)}

\item{constrain}{a logical, if TRUE then set all negative values to zero}
}
\description{
This takes a model and uses the XH component to define
the eirpr relationship using `xde_scaling_eir` then calls `xde_scaling_lambda`
}
