% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{pr2Lambda_history}
\alias{pr2Lambda_history}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
pr2Lambda_history(pr_ts, times, model, twice = TRUE)
}
\arguments{
\item{pr_ts}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` model}

\item{twice}{cycle through a second time}
}
\value{
an `xds` object
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
\note{
This utility relies on `xds_cohort`
}
