% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-pr2shockfit.R
\name{pr2shockfit}
\alias{pr2shockfit}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
pr2shockfit(xds_obj, fit_method = NULL)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{fit_method}{the method for \link{optim}}
}
\value{
the \strong{\code{ramp.xds}} model object, fitted to a time series.
The state at the end is saved as \code{xds_obj$history}
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
