% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-pr2shockit.R
\name{pr2shockfit_xm}
\alias{pr2shockfit_xm}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
pr2shockfit_xm(
  xds_obj,
  rep = 1,
  trend_fix = c(),
  trend_x = 1,
  bednet_fix = c(),
  bednet_x = 0,
  irs_fix = c(),
  irs_x = 0
)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{rep}{number of times to repeat the fitting cycle}

\item{trend_fix}{the indices for trend events to fix}

\item{trend_x}{the fixed values}

\item{bednet_fix}{the indices for bednet events to fix}

\item{bednet_x}{the fixed values}

\item{irs_fix}{the indices for irs events to fix}

\item{irs_x}{the fixed values}
}
\value{
the \strong{\code{ramp.xds}} model object, fitted to a time series.
The state at the end is saved as \code{xds_obj$history}
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
