% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-pr2shockit.R
\name{pr2shockit_xm}
\alias{pr2shockit_xm}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
pr2shockit_xm(xds_obj, bednet_ix = c(), irs_ix = c(), N = 1)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{bednet_ix}{the indices for bednet events to fit}

\item{irs_ix}{the indices for irs events to fit}

\item{N}{cycle through a N times}
}
\value{
the \strong{\code{ramp.xds}} model object, fitted to a time series.
The state at the end is saved as \code{xds_obj$history}
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
