% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2eir_history.R
\name{pr_ts2eir_history}
\alias{pr_ts2eir_history}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
pr_ts2eir_history(pr_ts, times, model)
}
\arguments{
\item{pr_ts}{the PR observed}

\item{times}{the times of the observations}

\item{model}{an `xds` model}
}
\value{
an `xds` object
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i


For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
\note{
This utility relies on `xds_cohort`
}
