% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-recon.R
\name{reconD_shock_xm}
\alias{reconD_shock_xm}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
reconD_shock_xm(
  xds_obj,
  do_bednet = TRUE,
  bednet_ix = c(),
  do_irs = TRUE,
  irs_ix = c(),
  D = 365,
  N = 2
)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{do_bednet}{if TRUE then fit bednet shocks}

\item{bednet_ix}{indices of events to fit}

\item{do_irs}{if TRUE then fit irs shocks}

\item{irs_ix}{indices of events to fit}

\item{D}{remove spline points less than D days after an event}

\item{N}{cycle through fits N times}
}
\value{
the \strong{\code{ramp.xds}} model object, fitted to a time series.
The state at the end is saved as \code{xds_obj$history}
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
