% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{reconstruct_baseline}
\alias{reconstruct_baseline}
\title{Reconstruct History as a Baseline Modified by Control}
\usage{
reconstruct_baseline(
  model,
  data,
  times,
  irs_rounds,
  bednets_rounds,
  impute_by = "mean"
)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{irs_rounds}{a list to set up multiround irs coverage}

\item{bednets_rounds}{a list to set up multiround bednets coverage}

\item{impute_by}{a string to dispatch the imputation method}
}
\description{
This function fits effect sizes for vector-based interventions. Each
intervention in a list includes a temporal profile describing how the
*effect* of the intervention wanes over time and the inception time.
Each model of intervention -- IRS or ITNs -- has its own sub-model that
describes how bionomic parameters are modified by vector control, given
contact and the waning effects.
}
