% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-pr2shockit.R
\name{restore_pr2shockit}
\alias{restore_pr2shockit}
\title{Reconstruct a history of exposure from a PR time series}
\usage{
restore_pr2shockit(xds_obj)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}
}
\value{
an \code{xds} object
}
\description{
Construct a function describing the EIR, including
the mean EIR, the seasonal pattern, and the i

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
\note{
This utility relies on \code{xds_scaling}
}
