% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{rm_ix_fit_spline_ty}
\alias{rm_ix_fit_spline_ty}
\title{Change Spline \eqn{t,y} Values}
\usage{
rm_ix_fit_spline_ty(ix, xds_obj)
}
\arguments{
\item{ix}{the indices of the values to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Change Spline \eqn{t,y} Values
}
