% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-setup.R
\name{scaling_init_ty.eir}
\alias{scaling_init_ty.eir}
\title{Use scaling to set initial guesses}
\usage{
\method{scaling_init_ty}{eir}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
For a xds_obj forced by emergence,
pre-calibrate the forcing function so that it's
close.
}
