% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fitting.R
\name{scaling_init_ty.Lambda}
\alias{scaling_init_ty.Lambda}
\title{Use scaling to set initial guesses}
\usage{
\method{scaling_init_ty}{Lambda}(model, pfpr, jdates)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{pfpr}{a *Pf*PR time series}

\item{jdates}{julian dates for `pfpr`}
}
\value{
a **`ramp.xds`** model object
}
\description{
For a model forced by emergence,
pre-calibrate the forcing function so that it's
close.
}
