% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fitting.R
\name{scaling_init_ty}
\alias{scaling_init_ty}
\title{Use scaling to set initial guesses}
\usage{
scaling_init_ty(model, pfpr, jdates)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{pfpr}{a *Pf*PR time series}

\item{jdates}{julian dates for `pfpr`}
}
\value{
a **`ramp.xds`** model object
}
\description{
Before fitting,
pre-calibrate the forcing function so that it's
close.
}
