% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast_ty.R
\name{set_pre_obs_y.mirror}
\alias{set_pre_obs_y.mirror}
\title{Set up the pre-observation interpolating points}
\usage{
\method{set_pre_obs_y}{mirror}(y, model)
}
\arguments{
\item{y}{the y values for interpolation}

\item{model}{a **`ramp.xds** model object}
}
\value{
a **`ramp.xds** model object
}
\description{
Set up the pre-observation interpolating points
}
\note{
Use "all" as the include_rule
}
