% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{set_pre_obs_y}
\alias{set_pre_obs_y}
\title{Set up a hindcast for burn-in}
\usage{
set_pre_obs_y(y, model)
}
\arguments{
\item{y}{the y values for interpolation}

\item{model}{a \strong{`ramp.xds} model object}
}
\value{
a \strong{`ramp.xds} model object
}
\description{
Set up a hindcast for burn-in
}
