% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_events.R
\name{setup_bednet_events}
\alias{setup_bednet_events}
\title{Set Up Bed Net Evaluation}
\usage{
setup_bednet_events(xds_obj, jdates, net_type, peak_access)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{jdates}{the julian dates of bed net mass distribution events}

\item{net_type}{the type of net used}

\item{peak_access}{the fraction of the population with access to a bed net}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Set Up Bed Net Evaluation
}
