% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-setup.R
\name{setup_data}
\alias{setup_data}
\title{Set up the data object for fitting}
\usage{
setup_data(xds_obj, pfpr, jdates, yr0 = 2015)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{pfpr}{a \emph{Pf}PR time series}

\item{jdates}{julian dates for \code{pfpr}}

\item{yr0}{the starting year}
}
\value{
an \strong{\code{xds_obj}}
}
\description{
Set up the data object for fitting
}
