% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet-dshape.R
\name{setup_fitting_indices.bednet_dshape}
\alias{setup_fitting_indices.bednet_dshape}
\title{Setup indices for bednet dshape}
\usage{
\method{setup_fitting_indices}{bednet_dshape}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
indices
}
\description{
Setup indices for bednet dshape
}
