% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_shock-size.R
\name{setup_fitting_indices.bednet_shock_size}
\alias{setup_fitting_indices.bednet_shock_size}
\title{Setup indices for bednet shock_size}
\usage{
\method{setup_fitting_indices}{bednet_shock_size}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
indices
}
\description{
Setup indices for bednet shock_size
}
