% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs-d50.R
\name{setup_fitting_indices.irs_d50}
\alias{setup_fitting_indices.irs_d50}
\title{Setup indices for irs shock_d50}
\usage{
\method{setup_fitting_indices}{irs_d50}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
indices
}
\description{
Setup indices for irs shock_d50
}
