% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs_shock.R
\name{setup_fitting_indices.irs_shock}
\alias{setup_fitting_indices.irs_shock}
\title{Setup indices for irs shock}
\usage{
\method{setup_fitting_indices}{irs_shock}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
indices
}
\description{
Setup indices for irs shock
}
