% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-mean_forcing.R
\name{setup_fitting_indices.mean_forcing}
\alias{setup_fitting_indices.mean_forcing}
\title{Indices for Mean Forcing}
\usage{
\method{setup_fitting_indices}{mean_forcing}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
options
}
\description{
Indices for Mean Forcing
}
