% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-season-phase.R
\name{setup_fitting_indices.phase}
\alias{setup_fitting_indices.phase}
\title{Seasonality: \code{phase} Indices}
\usage{
\method{setup_fitting_indices}{phase}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
options
}
\description{
Seasonality: \code{phase} Indices
}
