% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{setup_fitting_indices}
\alias{setup_fitting_indices}
\title{Set up indices for model fitting}
\usage{
setup_fitting_indices(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
options
}
\description{
Set up indices for model fitting
}
