% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season.R
\name{setup_fitting_indices.season}
\alias{setup_fitting_indices.season}
\title{Seasonality Indices}
\usage{
\method{setup_fitting_indices}{season}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
options
}
\description{
Seasonality Indices
}
