% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_setup.R
\name{setup_fitting_ts}
\alias{setup_fitting_ts}
\title{Set up the ts_obj fitting object}
\usage{
setup_fitting_ts(
  ts_obj,
  pfpr,
  jdates,
  yr0 = 2015,
  t_neg_inf = -365,
  N = 1,
  gof_method = "ts_sse"
)
}
\arguments{
\item{ts_obj}{a \strong{\code{ramp.xds}}  model object}

\item{pfpr}{a \emph{Pf}PR time series}

\item{jdates}{julian dates for \code{pfpr}}

\item{yr0}{the starting year}

\item{t_neg_inf}{a starting year}

\item{N}{the number of interpolation points for hindcasting and forecasting}

\item{gof_method}{to dispatch \link{compute_gof}}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Set up the ts_obj fitting object
}
