% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ty.R
\name{setup_forecast}
\alias{setup_forecast}
\title{Set the forecast interpolation}
\usage{
setup_forecast(
  model,
  N = 0,
  method = "value",
  impute_ty = "last",
  trust_ty = "last"
)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{N}{the number of years to forecast}

\item{method}{to dispatch set_post_obs_y}

\item{impute_ty}{to dispatch `impute_baseline_ty`}

\item{trust_ty}{to dispatch `get_trusted_ty`}
}
\value{
a **`ramp.xds`** model object
}
\description{
Set the forecast interpolation
}
