% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{setup_forecast_y.asis}
\alias{setup_forecast_y.asis}
\title{Set forecast interpolation points}
\usage{
\method{setup_forecast_y}{asis}(y, xds_obj)
}
\arguments{
\item{y}{the interpolation points}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
Set forecast interpolation points
}
