% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast_ty.R
\name{setup_hindcast}
\alias{setup_hindcast}
\title{Hindcast a Baseline}
\usage{
setup_hindcast(
  model,
  N = 0,
  method = "value",
  impute_ty = "mean",
  trust_ty = "first"
)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{N}{the number of years to forecast}

\item{method}{to dispatch set_pre_obs_ty}

\item{impute_ty}{to dispatch `impute_baseline_ty`}

\item{trust_ty}{to dispatch `get_trust_ty`}
}
\value{
a **`ramp.xds`** model object
}
\description{
This function adds N interpolation
points to the spline parameters *before* an
observation period for burn-in. The \eqn{t}
values are set at yearly intervals, and the
\eqn{y} values that serve as a basis. See
\href{https://dd-harp.github.io/ramp.work/vignettes/Hindcast.html}{Hindcasting}
}
