% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{setup_hindcast_y.mirror}
\alias{setup_hindcast_y.mirror}
\title{Set up the pre-observation interpolating points}
\usage{
\method{setup_hindcast_y}{mirror}(y, xds_obj)
}
\arguments{
\item{y}{the y values for interpolation}

\item{xds_obj}{a \strong{`ramp.xds} xds_obj object}
}
\value{
a \strong{`ramp.xds} xds_obj object
}
\description{
Set up the pre-observation interpolating points
}
\note{
Use "all" as the include_rule
}
