% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{setup_hindcast_y}
\alias{setup_hindcast_y}
\title{Set up a hindcast for burn-in}
\usage{
setup_hindcast_y(y, xds_obj)
}
\arguments{
\item{y}{the y values for interpolation}

\item{xds_obj}{a \strong{`ramp.xds} xds_obj object}
}
\value{
a \strong{`ramp.xds} xds_obj object
}
\description{
Set up a hindcast for burn-in
}
