% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{setup_imputation}
\alias{setup_imputation}
\title{Setup Imputation}
\usage{
setup_imputation(impute_ty = "mean", trusted_ty = "unmodified")
}
\arguments{
\item{impute_ty}{a text string to dispatch \code{impute_baseline_ty}}

\item{trusted_ty}{a text string to dispatch \code{get_trusted_ty}}
}
\value{
an imputation model object
}
\description{
This sets up the the imputatation model object. It
sets options for using interpolation points to impute a baseline
or to hindcast or forecast.
}
