% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_events.R
\name{setup_irs_events}
\alias{setup_irs_events}
\title{Set Up IRS Evaluation}
\usage{
setup_irs_events(xds_obj, jdates, pesticides, frac_sprayed)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{jdates}{the julian dates of IRS events}

\item{pesticides}{the pesticide used}

\item{frac_sprayed}{the fraction of houses sprayed}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Set Up IRS Evaluation
}
