% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{setup_pre_obs_y.value}
\alias{setup_pre_obs_y.value}
\title{Set up the pre-observation interpolating points}
\usage{
\method{setup_pre_obs_y}{value}(y, xds_obj)
}
\arguments{
\item{y}{the y values for interpolation}

\item{xds_obj}{a \strong{`ramp.xds} xds_obj object}
}
\value{
a \strong{`ramp.xds} xds_obj object
}
\description{
Set up the pre-observation interpolating points
}
\note{
Set the \eqn{y} value for
pre-observation interpolation points
to a constant.
}
