% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_setup.R
\name{setup_season.list}
\alias{setup_season.list}
\title{Set up the seasonal pattern}
\usage{
\method{setup_season}{list}(season_par)
}
\arguments{
\item{season_par}{seasonal pattern parameters}
}
\value{
a \code{sin} function parameter set
}
\description{
If the current \code{season_par} is a list,
return a parameter set
}
\seealso{
\code{\link[=makepar_F_sin]{makepar_F_sin()}}
}
