% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_setup.R
\name{setup_season.sin}
\alias{setup_season.sin}
\title{Set up the seasonal pattern}
\usage{
\method{setup_season}{sin}(season_par)
}
\arguments{
\item{season_par}{seasonal pattern parameters}
}
\value{
a \code{sin} function parameter set
}
\description{
If the current \code{season_par} is from the \code{sin}
family of functions, return the parameters unchanged
}
\seealso{
\code{\link[=makepar_F_sin]{makepar_F_sin()}}
}
