% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-setup.R
\name{setup_trend_par.eir}
\alias{setup_trend_par.eir}
\title{Check the trend setup}
\usage{
\method{setup_trend_par}{eir}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Check the trend setup
}
