% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_setup.R
\name{setup_trend_par}
\alias{setup_trend_par}
\title{Check the trend setup}
\usage{
setup_trend_par(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Check the trend setup
}
