% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vector_control_evaluation.R
\name{setup_vc_eval}
\alias{setup_vc_eval}
\title{Setup Vector Control for Evaluation}
\usage{
setup_vc_eval(model, irs_history, bednet_history)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{irs_history}{history of irs: jdate of spray round; insecticide name}

\item{bednet_history}{history of bednets: jdate of mass distribution; type bednets}
}
\value{
a **`ramp.xds`** model object
}
\description{
Setup Vector Control for Evaluation
}
