% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_events.R
\name{show_bednet_events}
\alias{show_bednet_events}
\title{Show bednet Events}
\usage{
show_bednet_events(xds_obj, mn = 0, mx = 1, clr = "#E4460AFF", add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{mn}{the bottom of the segment}

\item{mx}{the top of the segment}

\item{clr}{a color for the line segments}

\item{add}{if TRUE, add to an existing plot}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Show bednet Events
}
