% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_events.R
\name{show_events}
\alias{show_events}
\title{Show Events}
\usage{
show_events(
  xds_obj,
  mn = 0,
  mx = 1,
  bclr = "#E4460AFF",
  iclr = "#4686FBFF",
  add = FALSE
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{mn}{the bottom of the segment}

\item{mx}{the top of the segment}

\item{bclr}{a color for the bednet line segments}

\item{iclr}{a color for the irs line segments}

\item{add}{if TRUE, add to an existing plot}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Show Events
}
