% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{show_fit}
\alias{show_fit}
\title{Plot the model and the data}
\usage{
show_fit(xds_obj, clr = "black", add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clr}{the color of the line to be plotted}

\item{add}{add to an existing show_fit plot}
}
\value{
the PR, invisibly
}
\description{
Plot the model and the data
}
