% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{sigX}
\alias{sigX}
\title{\code{sigX}}
\usage{
sigX(x, mn = 0, mx = 1)
}
\arguments{
\item{x}{a number}

\item{mn}{the asymptotic value at negative infinity}

\item{mx}{the asymptotic value at infinity}
}
\value{
a number
}
\description{
Ensure the pw
parameter is properly bounded
}
