% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-compute_gof.R
\name{smooth_pr}
\alias{smooth_pr}
\title{Smooth PR}
\usage{
smooth_pr(pr, jdates, bwidth = 180)
}
\arguments{
\item{pr}{predicted PR values}

\item{jdates}{the julian dates}

\item{bwidth}{a constant}
}
\value{
smoothed PR
}
\description{
Smooth the PR
}
