% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X-xde-SIS.R
\name{split_stratum_by_biting.SIS}
\alias{split_stratum_by_biting.SIS}
\title{Split a stratum into two strata, assigning a fraction `p` a new biting weight that is multiplied by a factor `fac` compared with the old one. The biting weight for the remaining `1-p` gets a new factor `1/fac`}
\usage{
\method{split_stratum_by_biting}{SIS}(pars, i, j, p, fac)
}
\arguments{
\item{i}{the host species index}

\item{j}{the stratum to split}

\item{p}{the fraction that gets multiplied by `fac`}

\item{fac}{a factor}
}
\value{
pars a list

a [list]
}
\description{
Split a stratum into two strata, assigning a fraction `p` a new biting weight that is multiplied by a factor `fac` compared with the old one. The biting weight for the remaining `1-p` gets a new factor `1/fac`
}
