% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{sse_effect_size}
\alias{sse_effect_size}
\title{Given data, compute GoF for a spline function}
\usage{
sse_effect_size(X, data, times, model, event)
}
\arguments{
\item{X}{the value to be tested}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{the model}

\item{event}{relevant details of an event to assess}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series c(`times`,`data`),
compute the sum of squared errors for a seasonal
pattern defined by `Fpar
}
