% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{sse_Lambda_season}
\alias{sse_Lambda_season}
\title{Given data, compute GoF for a seasonal function}
\usage{
sse_Lambda_season(data, times, model)
}
\arguments{
\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{the model}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series c(`times`,`data`),
and a model, compute the sum of squared errors
}
