% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{sse_season}
\alias{sse_season}
\title{Given data, compute GoF for a seasonal function}
\usage{
sse_season(Fpar, data, times, model)
}
\arguments{
\item{Fpar}{parameters defining a seasonal pattern function}

\item{data}{the PR observed}

\item{times}{the times of the observations}

\item{model}{the model}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series c(`times`,`data`),
compute the sum of squared errors for a seasonal
pattern defined by `Fpar`
}
