% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xde.R
\name{ssMYZ}
\alias{ssMYZ}
\title{Set up the MYZss object for \code{xds_scaling_lambda}}
\usage{
ssMYZ(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\description{
This function computes several quantities that are require
}
