% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{time_since_event}
\alias{time_since_event}
\title{Time Since Event}
\usage{
time_since_event(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
numeric
}
\description{
Compute the time elapsed
since the last vector control event
for all the spline \eqn{t}-values
}
