% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_setup.R
\name{ts_setup}
\alias{ts_setup}
\title{Build a Decomposable Time Series Model}
\usage{
ts_setup(
  pr = 0.3,
  season_par = makepar_F_c(1),
  trend_par = makepar_F_c(1),
  age_par = makepar_F_c(1),
  shock_par = makepar_F_c(1),
  tnorm = 365
)
}
\arguments{
\item{pr}{is the mean pr}

\item{season_par}{parameters to configure a seasonality function using \link[ramp.func:make_function]{ramp.func::make_function}}

\item{trend_par}{parameters to configure a trends function using \link[ramp.func:make_function]{ramp.func::make_function}}

\item{age_par}{parameters to configure an age weights function using \link[ramp.func:make_function]{ramp.func::make_function}}

\item{shock_par}{parameters to configure a shock using \link[ramp.func:make_function]{ramp.func::make_function}}

\item{tnorm}{normalize from 0 up to tnorm}
}
\value{
a \strong{\code{ts}} object
}
\description{
Like \link[ramp.xds:xds_setup]{ramp.xds::xds_setup}, \link{ts_setup} sets up a decomposed,
multiplicative time-series analysis that can be used
to analyze malaria \emph{Pf}PR time series data in the
same way as an  \strong{\code{xds}} object.

The \strong{\code{ts}} object defines \code{frame = class(frame) = 'ts'}

The interface includes options to configure a function
describing \code{F_pr} as a function of time, with seasonal components
and a trend.

This can be used to model a cohort as it ages;
a function is set up to modify exposure by age.
}
