% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\name{UnevenAgingMatrix}
\alias{UnevenAgingMatrix}
\title{Create a cohort matrix}
\usage{
UnevenAgingMatrix(deathrates, ageMesh)
}
\arguments{
\item{deathrates}{the death rates for}

\item{ageMesh}{a partition on human age}
}
\value{
the steady states as a named vector
}
\description{
Return a matrix `dH`
}
