% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-mask_data.R
\name{unmask_data}
\alias{unmask_data}
\title{Unmask data}
\usage{
unmask_data(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
an \strong{\code{xds_obj}}
}
\description{
Restore the full data from the stored original
}
