% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-dts-step-ahead.R
\name{update_by_ar}
\alias{update_by_ar}
\title{Compute the next y}
\usage{
update_by_ar(ar, model, n = 365)
}
\arguments{
\item{ar}{the attack rate}

\item{model}{a model}

\item{n}{the number of times to iterate}
}
\value{
a vector of state variables, y
}
\description{
Compute the sum of squared errors
}
