% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{update_F_trend.eir}
\alias{update_F_trend.eir}
\title{Update `F_trend`}
\usage{
\method{update_F_trend}{eir}(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Use the `fitting` object to update `trend_par` and
`F_trend` when the model is forced by the eir
}
