% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{update_F_trend}
\alias{update_F_trend}
\title{Update `F_trend`}
\usage{
update_F_trend(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Use the `fitting` interpolation points to
update `trend_par` and `F_trend`
}
