% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend.R
\name{update_fit_trend}
\alias{update_fit_trend}
\title{feature Interpolation Points}
\usage{
update_fit_trend(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
This function copies the interpolation points from the
\code{hindcast} and \code{data_obj} and \code{forecast} to \code{fit_obj.}

During fit_obj, the algorithm changes
one or more of the \eqn{y} values of the
interpolation points in
\code{xds_obj$data_obj$yy.} The xds_obj then calls
\code{hindcast_ty} and \code{forecast_ty} to feature \code{xds_obj$hindcast$y} (the
pre-observation interpolation points for burn-in) and \code{xds_obj$hindcast$y}
(post-observation interpolation points).
These are copied to the
full set of control points stored at
at \code{xds_obj$fit_obj$tt} and \code{xds_obj$fit_obj$yy.}

The rules for \code{hindcast_ty}  and \code{forecast_ty}
are setup
by \code{setup_hindcast} and \code{setup_forecast.}
}
