% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{update_fitting_ty}
\alias{update_fitting_ty}
\title{feature Interpolation Points}
\usage{
update_fitting_ty(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
This function copies the interpolation points from the
\code{hindcast} and \code{data} and \code{forecast} to \code{fitting.}

During fitting, the algorithm changes
one or more of the \eqn{y} values of the
interpolation points in
\code{xds_obj$data$yy.} The xds_obj then calls
\code{hindcast_ty} and \code{forecast_ty} to feature \code{xds_obj$hindcast$y} (the
pre-observation interpolation points for burn-in) and \code{xds_obj$hindcast$y}
(post-observation interpolation points).
These are copied to the
full set of control points stored at
at \code{xds_obj$fitting$tt} and \code{xds_obj$fitting$yy.}

The rules for \code{hindcast_ty}  and \code{forecast_ty}
are setup
by \code{setup_hindcast} and \code{setup_forecast.}
}
