% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_amplitude.R
\name{update_function_X.amplitude}
\alias{update_function_X.amplitude}
\title{Update a function}
\usage{
\method{update_function_X}{amplitude}(X, xds_obj, update = "amplitude", ix = 1)
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{update}{a string to dispatch}

\item{ix}{indices or a list}
}
\value{
sum of squared differences
}
\description{
Update a function
}
