% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bednet_coverage.R
\name{update_function_X.bednet_coverage}
\alias{update_function_X.bednet_coverage}
\title{Update the bed net coverage function}
\usage{
\method{update_function_X}{bednet_coverage}(X, xds_obj, update = "bednet_coverage", ix = 1)
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{update}{a string to dispatch}

\item{ix}{indices or a list}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Update the bed net coverage function
}
