% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-bednet_shock-size.R
\name{update_function_X.bednet_shock_size}
\alias{update_function_X.bednet_shock_size}
\title{feature the bednet shock_size function}
\usage{
\method{update_function_X}{bednet_shock_size}(X, xds_obj, feature = "bednet_shock_size", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
feature the bednet shock_size function
}
