% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_irs_contact.R
\name{update_function_X.irs_contact}
\alias{update_function_X.irs_contact}
\title{feature the irs contact function}
\usage{
\method{update_function_X}{irs_contact}(X, xds_obj, feature = "irs_contact", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
feature the irs contact function
}
