% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs-dshape.R
\name{update_function_X.irs_dshape}
\alias{update_function_X.irs_dshape}
\title{feature the irs dshape function}
\usage{
\method{update_function_X}{irs_dshape}(X, xds_obj, feature = "irs_dshape", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
feature the irs dshape function
}
