% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs_shock.R
\name{update_function_X.irs_shock}
\alias{update_function_X.irs_shock}
\title{feature the irs shock function}
\usage{
\method{update_function_X}{irs_shock}(X, xds_obj, feature = "irs_shock", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
feature the irs shock function
}
