% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-mean_forcing.R
\name{update_function_X.mean_forcing}
\alias{update_function_X.mean_forcing}
\title{Update Mean Forcing}
\usage{
\method{update_function_X}{mean_forcing}(X, xds_obj, feature, options)
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
Update the mean value
that forces a xds_obj. This dispatches
on class(xds_obj$forced_by)
}
