% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{update_function_X}
\alias{update_function_X}
\title{Update a function}
\usage{
update_function_X(X, xds_obj, feature, options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
Update a function
}
