% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-season-bottom.R
\name{X2bottom}
\alias{X2bottom}
\title{\code{X2bottom}}
\usage{
X2bottom(x)
}
\arguments{
\item{x}{a number}
}
\value{
a number
}
\description{
Ensure the bottom
parameter is properly bounded
}
