% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-irs-dshape.R
\name{X2dshape}
\alias{X2dshape}
\title{\code{X2bottom}}
\usage{
X2dshape(x)
}
\arguments{
\item{x}{a number}
}
\value{
a number
}
\description{
Ensure the \code{d_shape}
parameter is properly bounded
}
