% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-conversion.R
\name{xde_eir2ni}
\alias{xde_eir2ni}
\title{Convert eir to ni}
\usage{
xde_eir2ni(eir, xds_obj, extend = FALSE)
}
\arguments{
\item{eir}{a \link{vector}}

\item{xds_obj}{a \link{list}}

\item{extend}{a \link{logical} option to determine whether to extend outside the range}
}
\description{
Use the scaling table to interpolate
}
