% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_terms.R
\name{xde_pr2Lambda}
\alias{xde_pr2Lambda}
\title{Convert pr to lambda}
\usage{
xde_pr2Lambda(pr, model, extend = FALSE)
}
\arguments{
\item{pr}{a [vector]}

\item{model}{a [list]}

\item{extend}{a [logical] option to determine whether to extend outside the range}
}
\description{
Use the outputs$lambdapr table to convert a set of pr values into lambda values
}
