% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_terms.R
\name{xde_pr2m}
\alias{xde_pr2m}
\title{Convert pr to mosquito density}
\usage{
xde_pr2m(pr, model, extend = FALSE)
}
\arguments{
\item{pr}{a [vector]}

\item{model}{a [list]}

\item{extend}{a [logical] option to determine whether to extend outside the range}
}
\description{
Use the outputs$eirpr table to convert a set of pr values into scaled mosquito density values
}
