% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xde.R
\name{xde_scaling.eir}
\alias{xde_scaling.eir}
\title{Compute eir-pr scaling relationships}
\usage{
\method{xde_scaling}{eir}(xds_obj, N = 25, rbr = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}

\item{rbr}{the relative biting rate for a population stratum}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This function calls \link{xds_solve_cohort} computes average annual values for the eir, the pr, and other
interesting terms and returns a table. It is computed for a xds_obj of class "cohort"
}
