% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xde.R
\name{xde_scaling.Lambda}
\alias{xde_scaling.Lambda}
\title{Compute scaling relationships from mosquito emergence through PfPR}
\usage{
\method{xde_scaling}{Lambda}(xds_obj, N = 30, rbr = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{N}{an integer}

\item{rbr}{the relative biting rate for a population stratum}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This function calls \link{xds_solve} to get the scaling
relationships for mosquito density over 9 factors of 10, and average annual
values for the eir, the pr, and other
interesting terms and returns a table.
}
