% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xde_scaling_eir}
\alias{xde_scaling_eir}
\title{Compute eir-pr scaling relationships}
\usage{
xde_scaling_eir(model, N = 25, rbr = 1)
}
\arguments{
\item{model}{a list that defines an `ramp.xds` model generated by `xde_setup_cohort()`}

\item{N}{an integer}

\item{rbr}{the relative biting rate for a population stratum}
}
\value{
**`xds`** model object
}
\description{
This function calls [xde_solve_cohort] computes average annual values for the eir, the pr, and other
interesting terms and returns a table. It is computed for a model of class "cohort"
}
