% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-plot.R
\name{xds_plot_eirpr}
\alias{xds_plot_eirpr}
\title{xds_plot the eir-pr scaling relationship}
\usage{
xds_plot_eirpr(xds_obj, clrs = "black", llty = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{a \link{list}}
}
\description{
xds_plot the eir-pr scaling relationship
}
