% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-plot.R
\name{xds_plot_eirVpr}
\alias{xds_plot_eirVpr}
\title{xds_plot EIR(t) \emph{vs.} the PR(t)}
\usage{
xds_plot_eirVpr(
  xds_obj,
  i = 1,
  clrs = "black",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{i}{the host species index}

\item{clrs}{color(s) for xds_plotting}

\item{llty}{an \link{integer} that specifies \code{lty} for xds_plotting}

\item{stable}{a \link{logical} set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a \link{logical} to xds_plot add_axes only if FALSE}
}
\description{
xds_plot EIR(t) \emph{vs.} the PR(t)
}
