% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xds.R
\name{xds_scaling.dts}
\alias{xds_scaling.dts}
\title{Scaling for Malaria Metrics}
\usage{
\method{xds_scaling}{dts}(xds_obj, N = 25, rbr = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}

\item{rbr}{the relative biting rate for a population stratum}
}
\value{
a \strong{\code{ramp.xds}} model object with a scaling table
}
\description{
For \code{class(xds) == dts} call \link{dts_scaling}
}
