% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xds_scaling}
\alias{xds_scaling}
\title{Scaling for Malaria Metrics}
\usage{
xds_scaling(xds_obj, N = 30)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Compute scaling relationships for malaria metrics
}
\note{
This function dispatches on \code{xds_obj$forced_by}
}
