% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xds_scaling_Lambda}
\alias{xds_scaling_Lambda}
\title{Compute}
\usage{
xds_scaling_Lambda(Lambda, xds_obj, scaling)
}
\arguments{
\item{Lambda}{emergence rate}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{scaling}{a list with stored values}
}
\value{
a pair of values
}
\description{
Compute
}
