% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling-plot.R
\name{add_orbits}
\alias{add_orbits}
\title{Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.}
\usage{
add_orbits(i, xds_obj, clr)
}
\arguments{
\item{i}{the index of the orbit to plot}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clr}{a \link{character} vector of colors}
}
\description{
Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling.
}
