% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{add_orbits_px}
\alias{add_orbits_px}
\title{Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling, and add points at the
minimum and maximum eir and pr}
\usage{
add_orbits_px(i, pars, clr)
}
\arguments{
\item{i}{the index of the orbit to plot}

\item{pars}{an **`xds`** object}

\item{clr}{a [character] vector of colors}
}
\description{
Draw the orbit for the \eqn{i^{th}} element of
eirpr$scaling, and add points at the
minimum and maximum eir and pr
}
