% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_compute_gof.R
\name{compute_gof}
\alias{compute_gof}
\title{Compute GoF}
\usage{
compute_gof(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
sum of squared differences
}
\description{
Compute GoF
}
