% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{compute_gof_X}
\alias{compute_gof_X}
\title{Compute the GoF for \code{X}}
\usage{
compute_gof_X(X, xds_obj, feature, options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{a string to dispatch a specific model feature}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
Compute the GoF for \code{X}
}
