% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_dispersal.R
\name{compute_IoD_F}
\alias{compute_IoD_F}
\title{Compute the Index of Dispersal for a function
F(t) over one year.}
\usage{
compute_IoD_F(F)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{clrs}{a [character] vector of colors}

\item{llty}{a [list]}
}
\description{
Compute the Index of Dispersal for a function
F(t) over one year.
}
