% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_dispersal.R
\name{compute_IoD_pr}
\alias{compute_IoD_pr}
\title{Compute the Index of Dispersal for a PR orbit
for the \eqn{i^{th}} element of
eirpr$scaling.}
\usage{
compute_IoD_pr(i, pars)
}
\arguments{
\item{pars}{an **`xds`** object}
}
\description{
Compute the Index of Dispersal for a PR orbit
for the \eqn{i^{th}} element of
eirpr$scaling.
}
