% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_of_dispersal.R
\name{compute_IoD_S}
\alias{compute_IoD_S}
\title{Compute the Index of Dispersal for a seasonal pattern
S(t) over one year. This assumes that
\deqn{\int_0^{365} S(t)dt = 365}
If in doubt, use [compute_IoD_F]}
\usage{
compute_IoD_S(S)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{clrs}{a [character] vector of colors}

\item{llty}{a [list]}
}
\description{
Compute the Index of Dispersal for a seasonal pattern
S(t) over one year. This assumes that
\deqn{\int_0^{365} S(t)dt = 365}
If in doubt, use [compute_IoD_F]
}
