% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xde.R
\name{compute_threshold}
\alias{compute_threshold}
\title{Get High/Low Values for Lambda}
\usage{
compute_threshold(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}
}
\value{
a pair of values
}
\description{
Get High/Low Values for Lambda
}
