% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mean_forcing.R
\name{fit_mean_forcing}
\alias{fit_mean_forcing}
\title{Fit mean forcing}
\usage{
fit_mean_forcing(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Fit the mean
forcing parameter. Functions
called by \code{fit_mean_forcing} dispatch
on \code{class(xds_obj$forced_by)}
\itemize{
\item \code{eir} updates the mean daily EIR
\item \code{Lambda} updates the mean daily emergence rate
}

The method is designed for fitting a model to
observed values of \emph{Pf}PR in a time series,
\eqn{x_t.}
}
\note{
The function parameter \code{options} is not used.
}
