% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-season.R
\name{fit_season}
\alias{fit_season}
\title{Fit a seasonal pattern}
\usage{
fit_season(xds_obj)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}
}
\value{
a list with the mean peak and the values
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
a spline function
}
