% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Fit interannual variability using splines}
\usage{
fit_trend(xds_obj, ix = c())
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}

\item{ix}{index of the \eqn{y} values of interpolation points}
}
\value{
a list with the mean peak and the values
}
\description{
For a \emph{Pf}PR time series, \eqn{x}, fit
a spline function
}
