% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Fit interannual variability using splines}
\usage{
fit_trend(pfpr_ts, jdates, model)
}
\arguments{
\item{pfpr_ts}{a PfPR time series}

\item{jdates}{the times of the observations}

\item{model}{an `xds` model}
}
\value{
a list with the mean peak and the values
}
\description{
For a *Pf*PR time series, \eqn{x}, fit
a spline function
}
