% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_spline_EIR}
\alias{forecast_spline_EIR}
\title{Add a "forecast"}
\usage{
forecast_spline_EIR(model, N = 5, x_last = 0)
}
\arguments{
\item{model}{a list that defines an `ramp.xds` model (*e.g.*,  generated by `xde_setup()`)}

\item{N}{the number of years ahead to forecast}

\item{x_last}{the number of last knot points to exclude}
}
\description{
Use the nodes of a trend function, defined as a spline, to
"forecast" the future by sub-sampling from the knots
}
