% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ty.R
\name{forecast_ty}
\alias{forecast_ty}
\title{Set the forecast interpolation}
\usage{
forecast_ty(model, ix = c())
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{ix}{(optional) indices for [impute_value]}
}
\value{
a **`ramp.xds`** model object
}
\description{
Set the forecast interpolation
}
