% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_ty}
\alias{forecast_ty}
\title{Set the forecast interpolation}
\usage{
forecast_ty(xds_obj, ix = c())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{ix}{(optional) indices for \link{impute_value}}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
Set the forecast interpolation
}
