% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_pw.R
\name{get_init_X.pw}
\alias{get_init_X.pw}
\title{Get Initial Values for Parameters}
\usage{
\method{get_init_X}{pw}(xds_obj, feature, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
