% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-compute_gof.R
\name{get_init_X}
\alias{get_init_X}
\title{Get Initial Values for Parameters}
\usage{
get_init_X(xds_obj, feature, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}

\item{options}{a list to configure feature-specific indexing}
}
\value{
a list with the mean peak and the values
}
\description{
Get Initial Values for Parameters
}
