% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mean_forcing.R
\name{get_limits_X.mean_forcing}
\alias{get_limits_X.mean_forcing}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{mean_forcing}(xds_obj, update = "mean_forcing")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{update}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
