% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_phase.R
\name{get_limits_X.phase}
\alias{get_limits_X.phase}
\title{Get Initial Values for Parameters}
\usage{
\method{get_limits_X}{phase}(xds_obj, feature = "phase")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{feature}{the dispatch string}
}
\value{
a vector
}
\description{
Get Initial Values for Parameters
}
