% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_baseline.R
\name{impute_baseline.median}
\alias{impute_baseline.median}
\title{Use Median for Modify Baseline}
\usage{
\method{impute_baseline}{median}(ix, model)
}
\arguments{
\item{ix}{the index of the node to replace}

\item{model}{a **`ramp.xds`** model}
}
\value{
an object to dispatch [modify_baseline.ttyy]
}
\description{
This function modifies
}
