% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_baseline_ty}
\alias{impute_baseline_ty}
\title{Impute the baseline}
\usage{
impute_baseline_ty(ix, xds_obj, impute_ty, trust_ty)
}
\arguments{
\item{ix}{the index (or indices) of a node (or nodes) to replace}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{impute_ty}{a text string to dispatch \code{impute_baseline_ty}}

\item{trust_ty}{a text string to dispatch \code{get_ty}}
}
\value{
numeric
}
\description{
Impute the \eqn{n} value of one or more
interpolation points in the pair, \eqn{t,y}.
}
