% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_value}
\alias{impute_value}
\title{Impute the baseline}
\usage{
impute_value(y, impute_y, N = 1)
}
\arguments{
\item{y}{a set of y values}

\item{impute_y}{a text string to dispatch \code{impute_value}}

\item{N}{the length of the return value}
}
\value{
numeric
}
\description{
Impute the \eqn{n} value of one or more
interpolation points in the pair, \eqn{t,y}.
}
