% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-season.R
\name{init_fit_season.list}
\alias{init_fit_season.list}
\title{Set up the seasonal pattern}
\usage{
\method{init_fit_season}{list}(season_par)
}
\arguments{
\item{season_par}{seasonal pattern parameters}
}
\value{
a \code{sin} function parameter set
}
\description{
If the current \code{season_par} is a list,
return a parameter set
}
\seealso{
\code{\link[ramp.xds:makepar_F_sin]{ramp.xds::makepar_F_sin()}}
}
