% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-season.R
\name{init_fit_season.sin}
\alias{init_fit_season.sin}
\title{Set up the seasonal pattern}
\usage{
\method{init_fit_season}{sin}(season_par)
}
\arguments{
\item{season_par}{seasonal pattern parameters}
}
\value{
a \code{sin} function parameter set
}
\description{
If the current \code{season_par} is from the \code{sin}
family of functions, return the parameters unchanged
}
\seealso{
\code{\link[ramp.xds:makepar_F_sin]{ramp.xds::makepar_F_sin()}}
}
