% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_plot.R
\name{lines_eirpr}
\alias{lines_eirpr}
\title{Add lines for an eir-pr scaling relationship}
\usage{
lines_eirpr(xds_obj, clrs = "black", llty = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{a \link{list}}
}
\description{
Add lines for an eir-pr scaling relationship
}
