% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{make_vc_event_list}
\alias{make_vc_event_list}
\title{Make an events list}
\usage{
make_vc_event_list(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a modified  **`ramp.xds`** model object
}
\description{
Return an ordered list of events by time and type
}
