% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-setup.R
\name{mask_data}
\alias{mask_data}
\title{Mask data}
\usage{
mask_data(xds_obj, mask_ix)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
an \strong{\code{xds_obj}}
}
\description{
Mask data
}
