% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-setup.R
\name{mask_events}
\alias{mask_events}
\title{Mask data}
\usage{
mask_events(xds_obj, daysb4 = 0, days_after = 365)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
an \strong{\code{xds_obj}}
}
\description{
Mask data
}
