% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{mean_phase_peak}
\alias{mean_phase_peak}
\title{Compute the phase of the peak}
\usage{
mean_phase_peak(t, X, window = 170)
}
\arguments{
\item{t}{the times}

\item{X}{the variable}

\item{window}{days around t}
}
\value{
a list with the mean peak and the values
}
\description{
For a time series of paired values \eqn{c(t,X)} compute the
phase, the time of the year when there is a peak
}
