% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_plot.R
\name{plot_eirpr}
\alias{plot_eirpr}
\title{Plot the eir-pr scaling relationship}
\usage{
plot_eirpr(xds_obj, clrs = "black", llty = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{a \link{list}}
}
\description{
Plot the eir-pr scaling relationship
}
