% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_eirVpr}
\alias{plot_eirVpr}
\title{Plot EIR(t) *vs.* the PR(t)}
\usage{
plot_eirVpr(
  pars,
  i = 1,
  clrs = "black",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a [list] that defines an `exDE` model (*e.g.*,  generated by `exDE::xde_setup()`)}

\item{i}{the host species index}

\item{clrs}{a [character] vector of colors}

\item{llty}{an [integer] that specifies `lty` for plotting}

\item{stable}{a [logical] set to FALSE for `orbits` and TRUE for `stable_orbits`}

\item{add_axes}{a [logical] to plot add_axes only if FALSE}
}
\description{
Plot EIR(t) *vs.* the PR(t)
}
