% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{pr2Lambda}
\alias{pr2Lambda}
\title{Using the eirpr matrix and a RM xds_obj, convert pr to Lambda}
\usage{
pr2Lambda(pr, xds_obj, constrain = TRUE)
}
\arguments{
\item{pr}{a vector}

\item{xds_obj}{a list that defines an \code{ramp.xds} xds_obj (\emph{e.g.},  generated by \code{xde_setup()})}

\item{constrain}{a logical, if TRUE then set all negative values to zero}
}
\description{
This takes a xds_obj and uses the XH component to define
the eirpr relationship using \code{xde_scaling_eir} then calls \code{xde_scaling_lambda}
}
