% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xde.R
\name{pr2Lambda}
\alias{pr2Lambda}
\title{Using the eirpr matrix and a RM xds_obj, convert pr to Lambda}
\usage{
pr2Lambda(pr, xds_obj, constrain = TRUE)
}
\arguments{
\item{pr}{a vector}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{constrain}{a logical, if TRUE then set all negative values to zero}
}
\description{
This takes a xds_obj and uses the XH component to define
the eirpr relationship using \code{xds_scaling} then calls \code{xds_scaling}
}
