% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season_phase.R
\name{preset_phase}
\alias{preset_phase}
\title{Initialize the phase parameter}
\usage{
preset_phase(xds_obj)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}
}
\value{
a list with the mean peak and the values
}
\description{
Compute the empirical phase
from a \emph{Pf}PR time series, \eqn{t,x}, and
adjust the model phase to match it.
}
\note{
This algorithm makes a crude guess
at the seasonal phase for the forcing function.
\emph{close} to the value that
will get fitted later.
}
