% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_baseline.R
\name{reconstruct_baseline}
\alias{reconstruct_baseline}
\title{Reconstruct History as a Baseline Modified by Control}
\usage{
reconstruct_baseline(
  model,
  pfpr_ts,
  jdates,
  irs_history,
  bednet_history,
  impute_ty = "mean",
  trust_ty = "unmodified"
)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{pfpr_ts}{the *Pf*PR time series}

\item{jdates}{the times of the observations}

\item{irs_history}{history of irs: jdate of spray round; insecticide name}

\item{bednet_history}{history of bednets: jdate of mass distribution; type bednets}

\item{impute_ty}{a text string to dispatch `impute_baseline_ty`}

\item{trust_ty}{a text string to dispatch `get_trust_ty`}
}
\description{
This function fits effect sizes for vector-based interventions. Each
intervention in a list includes a temporal profile describing how the
*effect* of the intervention wanes over time and the inception time.
Each model of intervention -- IRS or ITNs -- has its own sub-model that
describes how bionomic parameters are modified by vector control, given
contact and the waning effects.
}
