% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr2history.R
\name{save_pr2history}
\alias{save_pr2history}
\title{Save the history fitting}
\usage{
save_pr2history(xds_obj)
}
\arguments{
\item{xds_obj}{an \code{xds} xds_obj}
}
\value{
an \code{xds} object
}
\description{
After fitting a model
to data, save the history. To restore
that history, use []

For set of paired a time series \eqn{X,} compute the
phase of a seasonal pattern for the EIR
}
\note{
This utility relies on \code{xds_cohort}
}
