% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-setup.R
\name{scaling_init_ty}
\alias{scaling_init_ty}
\title{Use scaling to set initial guesses}
\usage{
scaling_init_ty(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Before fitting,
pre-calibrate the forcing function so that it's
close.
}
