% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ty.R
\name{set_post_obs_y.asis}
\alias{set_post_obs_y.asis}
\title{Set forecast interpolation points}
\usage{
\method{set_post_obs_y}{asis}(y, model)
}
\arguments{
\item{y}{the interpolation points}

\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
Set forecast interpolation points
}
