% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{set_post_obs_y}
\alias{set_post_obs_y}
\title{Forecast a Baseline}
\usage{
set_post_obs_y(y, model)
}
\arguments{
\item{y}{the interpolation points}

\item{model}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Forecast a Baseline
}
