% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast_ty.R
\name{set_pre_obs_y.value}
\alias{set_pre_obs_y.value}
\title{Set up the pre-observation interpolating points}
\usage{
\method{set_pre_obs_y}{value}(y, model)
}
\arguments{
\item{y}{the y values for interpolation}

\item{model}{a **`ramp.xds** model object}
}
\value{
a **`ramp.xds** model object
}
\description{
Set up the pre-observation interpolating points
}
\note{
Set the \eqn{y} value for
pre-observation interpolation points
to a constant.
}
