% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_setup.R
\name{setup_fitting}
\alias{setup_fitting}
\title{Set up the xds_obj fitting object}
\usage{
setup_fitting(
  xds_obj,
  pfpr,
  jdates,
  t_neg_inf = 3650,
  N = 10,
  pr_diagnostic = "rdt",
  gof_method = "sse"
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{pfpr}{a \emph{Pf}PR time series}

\item{jdates}{julian dates for \code{pfpr}}

\item{t_neg_inf}{a date (in the past) to use for burnin}

\item{N}{the number of interpolation points for hindcasting and forecasting}

\item{pr_diagnostic}{the method used to estimate the \emph{Pf}PR data}

\item{gof_method}{to dispatch \link{compute_gof}}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Set up the xds_obj fitting object
}
