% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fitting.R
\name{setup_fitting}
\alias{setup_fitting}
\title{Set up the model fitting object}
\usage{
setup_fitting(
  model,
  pfpr,
  jdates,
  N = 10,
  forced_by = "Lambda",
  pr_diagnostic = "rdt",
  trend_par = "spline",
  gof = "sse"
)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{pfpr}{a *Pf*PR time series}

\item{jdates}{julian dates for `pfpr`}

\item{N}{the number of interpolation points for hindcasting and forecasting}

\item{forced_by}{emergence = Lambda; eir = eir;}

\item{pr_diagnostic}{the method used to estimate the *Pf*PR data}

\item{trend_par}{to dispatch [fit_trend]}

\item{gof}{to dispatch [compute_gof]}
}
\value{
a **`ramp.xds`** model object
}
\description{
Set up the model fitting object
}
