% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{setup_fitting_indices.trend}
\alias{setup_fitting_indices.trend}
\title{Setup indices for}
\usage{
\method{setup_fitting_indices}{trend}(xds_obj, feature, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{feature}{a list of model features to feature}

\item{options}{options for setting indices}
}
\value{
indices
}
\description{
Setup indices for
}
