% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{setup_forecast}
\alias{setup_forecast}
\title{Set the forecast interpolation}
\usage{
setup_forecast(xds_obj, N = 0, method = "use_last")
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{N}{the number of years to forecast}

\item{method}{to dispatch setup_forecast_y}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
Set the forecast interpolation
}
