% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{setup_forecast}
\alias{setup_forecast}
\title{Set the forecast interpolation}
\usage{
setup_forecast(
  model,
  N = 0,
  method = "value",
  impute_ty = "last",
  trust_ty = "last"
)
}
\arguments{
\item{model}{a \strong{\code{ramp.xds}} model object}

\item{N}{the number of years to forecast}

\item{method}{to dispatch setup_post_obs_y}

\item{impute_ty}{to dispatch \code{impute_baseline_ty}}

\item{trust_ty}{to dispatch \code{get_trusted_ty}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the forecast interpolation
}
