% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindcast.R
\name{setup_hindcast}
\alias{setup_hindcast}
\title{Hindcast a Baseline}
\usage{
setup_hindcast(
  xds_obj,
  N = 0,
  method = "value",
  impute_ty = "mean",
  trust_ty = "first"
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{N}{the number of years to hindcast}

\item{method}{to dispatch setup_pre_obs_ty}

\item{impute_ty}{to dispatch \code{impute_baseline_ty}}

\item{trust_ty}{to dispatch \code{get_trust_ty}}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
This function adds N interpolation
points to the spline parameters \emph{before} an
observation period for burn-in. The \eqn{t}
values are set at yearly intervals, and the
\eqn{y} values that serve as a basis. See
\href{https://dd-harp.github.io/ramp.work/vignettes/Hindcast.html}{Hindcasting}
}
