% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{setup_post_obs_y}
\alias{setup_post_obs_y}
\title{Forecast a Baseline}
\usage{
setup_post_obs_y(y, xds_obj)
}
\arguments{
\item{y}{the interpolation points}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
Forecast a Baseline
}
