% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting-setup.R
\name{setup_season}
\alias{setup_season}
\title{Set up the seasonal pattern}
\usage{
setup_season(season_par)
}
\arguments{
\item{season_par}{seasonal pattern parameters}
}
\value{
a \code{sin} function parameter set
}
\description{
Set up the seasonal pattern
}
\seealso{
\code{\link[ramp.xds:makepar_F_sin]{ramp.xds::makepar_F_sin()}}
}
