% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-trend-utils.R
\name{time_since_event}
\alias{time_since_event}
\title{Time Since Event}
\usage{
time_since_event(xds_obj, bednet_ix = list(), irs_ix = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{bednet_ix}{list of events to fit; if empty, then fit all}

\item{irs_ix}{list of events to fit; if empty, then fit all}
}
\value{
numeric
}
\description{
Compute the time elapsed
since the last vector control event
for all the spline \eqn{t}-values
}
