% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{update_fitting_ty}
\alias{update_fitting_ty}
\title{Update Interpolation Points}
\usage{
update_fitting_ty(model)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}
}
\value{
a **`ramp.xds`** model object
}
\description{
During fitting, the algorithm changes
one or more of the control points in
`model$fitting$data$yy.` The full set of control points
at `model$fitting$tt` and `model$fitting$yy` includes
some pre-observation interpolation points for burn-in, a set that
brackets the observation period, and post-observation
interpolation points to ensure continuity.

This function copies the interpolation points from the
separate sets to the master set
}
