% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-season-amplitude.R
\name{update_function_X.amplitude}
\alias{update_function_X.amplitude}
\title{feature a function}
\usage{
\method{update_function_X}{amplitude}(X, xds_obj, feature = "amplitude", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
feature a function
}
