% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bednet_coverage.R
\name{update_function_X.bednet_coverage}
\alias{update_function_X.bednet_coverage}
\title{feature the bed net coverage function}
\usage{
\method{update_function_X}{bednet_coverage}(X, xds_obj, feature = "bednet_coverage", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
feature the bed net coverage function
}
