% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_compute_gof.R
\name{update_function_X}
\alias{update_function_X}
\title{Update a function}
\usage{
update_function_X(X, xds_obj, update, ix = 1)
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{update}{a string to dispatch}

\item{ix}{indices or a list}
}
\value{
sum of squared differences
}
\description{
Update a function
}
