% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_season.R
\name{update_function_X.season}
\alias{update_function_X.season}
\title{feature \code{F_season}}
\usage{
\method{update_function_X}{season}(X, xds_obj, feature = "season", options = list())
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
This features three shape parameters
for a function \code{F_season} using \link{modify_vector_X} and \link[ramp.xds:update_F_season]{ramp.xds::update_F_season}
}
\note{
This assumes the vector, \code{X}, has got
three sets of parameters describing three
parameters in a seasonal function with
parameter names phase, bottom, and pw.
}
