% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-feature-trend.R
\name{update_function_X.trend}
\alias{update_function_X.trend}
\title{Update the Trend Function}
\usage{
\method{update_function_X}{trend}(X, xds_obj, feature, options)
}
\arguments{
\item{X}{new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{feature}{a string to dispatch}

\item{options}{a list to configure feature-specific indexing}
}
\value{
sum of squared differences
}
\description{
Update the Trend Function
}
