% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-conversion.R
\name{xde_pr2m}
\alias{xde_pr2m}
\title{Convert pr to mosquito density}
\usage{
xde_pr2m(pr, xds_obj, extend = FALSE)
}
\arguments{
\item{pr}{a \link{vector}}

\item{xds_obj}{a \link{list}}

\item{extend}{a \link{logical} option to determine whether to extend outside the range}
}
\description{
Use the scaling table to convert a set of pr values into scaled mosquito density values
}
