% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xde_scaling_lambda}
\alias{xde_scaling_lambda}
\title{Compute scaling relationships from mosquito emergence through PfPR}
\usage{
xde_scaling_lambda(model)

xde_scaling_lambda(model)
}
\arguments{
\item{model}{a list that defines an `ramp.xds` model (*e.g.*,  generated by `xds_setup()`)}

\item{N}{an integer}
}
\value{
**`xds`** model object
}
\description{
This function calls [xde_solve] to get the scaling
relationships for mosquito density over 9 factors of 10, and average annual
values for the eir, the pr, and other
interesting terms and returns a table.

This function computes `m` and `lambda` for the output of one
of `xde_scaling_eir` or `xde_scaling_Z`.  The outputs are attached to eirpr
}
