% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xde_scaling_lambda}
\alias{xde_scaling_lambda}
\title{Compute scaling relationships from mosquito emergence through PfPR}
\usage{
xde_scaling_lambda(model, fac = 1.2, N = 30)
}
\arguments{
\item{model}{a **`ramp.xds`** model object}

\item{fac}{a multiplicative factor that defines the mesh on \eqn{\log \Lambda}}

\item{N}{an integer}
}
\value{
**`xds`** model object
}
\description{
This function calls [xde_solve] to get the scaling
relationships for mosquito density over 9 factors of 10, and average annual
values for the eir, the pr, and other
interesting terms and returns a table.
}
