% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_xds.R
\name{xds_scaling}
\alias{xds_scaling}
\title{Scaling for Malaria Metrics}
\usage{
xds_scaling(xds_obj, N = 25, rbr = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}

\item{rbr}{the relative biting rate for a population stratum}
}
\value{
a \strong{\code{ramp.xds}} model object with a scaling table
}
\description{
Compute scaling relationships for malaria metrics

This function is a wrapper that dispatches on \code{xds} to call
\itemize{
\item \link{xde_scaling} for differential equations
\item \link{dts_scaling} for discrete-time systems
}
}
